Block sampling

Sample blockwise from clustered data

blocksample(data, size, idvar = NULL, replace = TRUE, ...)

Arguments

data
Data frame

size
Size of samples

idvar
Column defining the clusters

replace
Logical indicating wether to sample with replacement

...
additional arguments to lower level functions

Value

data.frame

Details

Original id is stored in the attribute 'id'

Examples

d <- data.frame(x=rnorm(5), z=rnorm(5), id=c(4,10,10,5,5), v=rnorm(5))
(dd <- blocksample(d,size=20,~id))

#>               x           z id          v
#> 2    -0.8201687  1.46936229  1 -0.5449456
#> 3     0.3127244 -0.80220037  1  0.9535810
#> 4    -1.0063361  0.05593903  2 -0.4298104
#> 5     0.1103187 -1.78243993  2  1.9748455
#> 4.1  -1.0063361  0.05593903  3 -0.4298104
#> 5.1   0.1103187 -1.78243993  3  1.9748455
#> 2.1  -0.8201687  1.46936229  4 -0.5449456
#> 3.1   0.3127244 -0.80220037  4  0.9535810
#> 2.2  -0.8201687  1.46936229  5 -0.5449456
#> 3.2   0.3127244 -0.80220037  5  0.9535810
#> 4.2  -1.0063361  0.05593903  6 -0.4298104
#> 5.2   0.1103187 -1.78243993  6  1.9748455
#> 1    -0.0762826 -0.18449948  7 -0.2344004
#> 4.3  -1.0063361  0.05593903  8 -0.4298104
#> 5.3   0.1103187 -1.78243993  8  1.9748455
#> 4.4  -1.0063361  0.05593903  9 -0.4298104
#> 5.4   0.1103187 -1.78243993  9  1.9748455
#> 4.5  -1.0063361  0.05593903 10 -0.4298104
#> 5.5   0.1103187 -1.78243993 10  1.9748455
#> 2.3  -0.8201687  1.46936229 11 -0.5449456
#> 3.3   0.3127244 -0.80220037 11  0.9535810
#> 2.4  -0.8201687  1.46936229 12 -0.5449456
#> 3.4   0.3127244 -0.80220037 12  0.9535810
#> 1.1  -0.0762826 -0.18449948 13 -0.2344004
#> 4.6  -1.0063361  0.05593903 14 -0.4298104
#> 5.6   0.1103187 -1.78243993 14  1.9748455
#> 4.7  -1.0063361  0.05593903 15 -0.4298104
#> 5.7   0.1103187 -1.78243993 15  1.9748455
#> 1.2  -0.0762826 -0.18449948 16 -0.2344004
#> 4.8  -1.0063361  0.05593903 17 -0.4298104
#> 5.8   0.1103187 -1.78243993 17  1.9748455
#> 2.5  -0.8201687  1.46936229 18 -0.5449456
#> 3.5   0.3127244 -0.80220037 18  0.9535810
#> 4.9  -1.0063361  0.05593903 19 -0.4298104
#> 5.9   0.1103187 -1.78243993 19  1.9748455
#> 4.10 -1.0063361  0.05593903 20 -0.4298104
#> 5.10  0.1103187 -1.78243993 20  1.9748455

attributes(dd)$id

#>  [1] 10 10  5  5  5  5 10 10 10 10  5  5  4  5  5  5  5  5  5 10 10 10 10  4  5
#> [26]  5  5  5  4  5  5 10 10  5  5  5  5

if (FALSE) {
blocksample(data.table::data.table(d),1e6,~id)
}

d <- data.frame(x=c(1,rnorm(9)),
               z=rnorm(10),
               id=c(4,10,10,5,5,4,4,5,10,5),
               id2=c(1,1,2,1,2,1,1,1,1,2),
               v=rnorm(10))
dsample(d,~id, size=2)

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

#>             x           z id id2          v id.1
#> 4   1.4675950  0.08727096  5   1  0.7719504    1
#> 5   0.7284417  0.22424031  5   2 -1.2617650    1
#> 8   0.1311181  0.53218413  5   1 -0.6092151    1
#> 10  0.7818185 -0.01969977  5   2 -0.4098266    1
#> 2  -3.0094030 -1.33914184 10   1 -1.3352669    2
#> 3   0.2706985 -0.02015390 10   2 -0.3334107    2
#> 9  -1.4614550  0.59899580 10   1  0.4814989    2

dsample(d,.~id+id2)

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

#>               x           z          v id
#> 5     0.7284417  0.22424031 -1.2617650  1
#> 10    0.7818185 -0.01969977 -0.4098266  1
#> 3     0.2706985 -0.02015390 -0.3334107  2
#> 1     1.0000000 -0.16231932 -0.6765979  3
#> 6    -0.7070281 -1.71510190 -0.1298113  3
#> 7     1.0952601 -0.86704863 -0.1509304  3
#> 5.1   0.7284417  0.22424031 -1.2617650  4
#> 10.1  0.7818185 -0.01969977 -0.4098266  4
#> 5.2   0.7284417  0.22424031 -1.2617650  5
#> 10.2  0.7818185 -0.01969977 -0.4098266  5
#> 4     1.4675950  0.08727096  0.7719504  6
#> 8     0.1311181  0.53218413 -0.6092151  6
#> 5.3   0.7284417  0.22424031 -1.2617650  7
#> 10.3  0.7818185 -0.01969977 -0.4098266  7
#> 1.1   1.0000000 -0.16231932 -0.6765979  8
#> 6.1  -0.7070281 -1.71510190 -0.1298113  8
#> 7.1   1.0952601 -0.86704863 -0.1509304  8
#> 4.1   1.4675950  0.08727096  0.7719504  9
#> 8.1   0.1311181  0.53218413 -0.6092151  9
#> 5.4   0.7284417  0.22424031 -1.2617650 10
#> 10.4  0.7818185 -0.01969977 -0.4098266 10

dsample(d,x+z~id|x>0,size=5)

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

#>              x           z id
#> 4    1.4675950  0.08727096  1
#> 5    0.7284417  0.22424031  1
#> 8    0.1311181  0.53218413  1
#> 10   0.7818185 -0.01969977  1
#> 1    1.0000000 -0.16231932  2
#> 7    1.0952601 -0.86704863  2
#> 4.1  1.4675950  0.08727096  3
#> 5.1  0.7284417  0.22424031  3
#> 8.1  0.1311181  0.53218413  3
#> 10.1 0.7818185 -0.01969977  3
#> 4.2  1.4675950  0.08727096  4
#> 5.2  0.7284417  0.22424031  4
#> 8.2  0.1311181  0.53218413  4
#> 10.2 0.7818185 -0.01969977  4
#> 1.1  1.0000000 -0.16231932  5
#> 7.1  1.0952601 -0.86704863  5