Calculate summary statistics grouped by

Calculate summary statistics grouped by variable

dby(
  data,
  INPUT,
  ...,
  ID = NULL,
  ORDER = NULL,
  SUBSET = NULL,
  SORT = 0,
  COMBINE = !REDUCE,
  NOCHECK = FALSE,
  ARGS = NULL,
  NAMES,
  COLUMN = FALSE,
  REDUCE = FALSE,
  REGEX = mets.options()$regex,
  ALL = TRUE
)

Arguments

data
Data.frame

INPUT
Input variables (character or formula)

...
functions

ID
id variable

ORDER
(optional) order variable

SUBSET
(optional) subset expression

SORT
sort order (id+order variable)

COMBINE
If TRUE result is appended to data

NOCHECK
No sorting or check for missing data

ARGS
Optional list of arguments to functions (...)

NAMES
Optional vector of column names

COLUMN
If TRUE do the calculations for each column

REDUCE
Reduce number of redundant rows

REGEX
Allow regular expressions

ALL
if FALSE only the subset will be returned

Details

Calculate summary statistics grouped by

dby2 for column-wise calculations

Examples

n <- 4
k <- c(3,rbinom(n-1,3,0.5)+1)
N <- sum(k)
d <- data.frame(y=rnorm(N),x=rnorm(N),id=rep(seq(n),k),num=unlist(sapply(k,seq)))
d2 <- d[sample(nrow(d)),]

dby(d2, y~id, mean)

#>              y           x id num        mean
#> 3  -1.55336807 -2.03545050  1   3 -1.06455561
#> 1  -1.79543538  0.55634005  1   1 -1.06455561
#> 2   0.15513661  0.03833857  1   2 -1.06455561
#> 4  -0.71603154 -0.06096876  2   1 -0.26978853
#> 5  -0.87809866  1.05140078  2   2 -0.26978853
#> 6   0.78476462 -0.11705507  2   3 -0.26978853
#> 10 -0.30420261  0.19521967  3   4 -0.03992063
#> 7   0.36620380 -1.21157397  3   1 -0.03992063
#> 8  -0.20842081  1.16618593  3   2 -0.03992063
#> 9  -0.01326291  0.46188895  3   3 -0.03992063
#> 11 -0.24722026 -0.18616474  4   1 -0.24722026

dby(d2, y~id + order(num), cumsum)

#>              y           x id num     cumsum
#> 1  -1.79543538  0.55634005  1   1 -1.7954354
#> 2   0.15513661  0.03833857  1   2 -1.6402988
#> 3  -1.55336807 -2.03545050  1   3 -3.1936668
#> 4  -0.71603154 -0.06096876  2   1 -0.7160315
#> 5  -0.87809866  1.05140078  2   2 -1.5941302
#> 6   0.78476462 -0.11705507  2   3 -0.8093656
#> 7   0.36620380 -1.21157397  3   1  0.3662038
#> 8  -0.20842081  1.16618593  3   2  0.1577830
#> 9  -0.01326291  0.46188895  3   3  0.1445201
#> 10 -0.30420261  0.19521967  3   4 -0.1596825
#> 11 -0.24722026 -0.18616474  4   1 -0.2472203

dby(d,y ~ id + order(num), dlag)

#>              y           x id num        dlag
#> 1  -1.79543538  0.55634005  1   1          NA
#> 2   0.15513661  0.03833857  1   2 -1.79543538
#> 3  -1.55336807 -2.03545050  1   3  0.15513661
#> 4  -0.71603154 -0.06096876  2   1          NA
#> 5  -0.87809866  1.05140078  2   2 -0.71603154
#> 6   0.78476462 -0.11705507  2   3 -0.87809866
#> 7   0.36620380 -1.21157397  3   1          NA
#> 8  -0.20842081  1.16618593  3   2  0.36620380
#> 9  -0.01326291  0.46188895  3   3 -0.20842081
#> 10 -0.30420261  0.19521967  3   4 -0.01326291
#> 11 -0.24722026 -0.18616474  4   1          NA

dby(d,y ~ id + order(num), dlag, ARGS=list(k=1:2))

#>              y           x id num       dlag1      dlag2
#> 1  -1.79543538  0.55634005  1   1          NA         NA
#> 2   0.15513661  0.03833857  1   2 -1.79543538         NA
#> 3  -1.55336807 -2.03545050  1   3  0.15513661 -1.7954354
#> 4  -0.71603154 -0.06096876  2   1          NA         NA
#> 5  -0.87809866  1.05140078  2   2 -0.71603154         NA
#> 6   0.78476462 -0.11705507  2   3 -0.87809866 -0.7160315
#> 7   0.36620380 -1.21157397  3   1          NA         NA
#> 8  -0.20842081  1.16618593  3   2  0.36620380         NA
#> 9  -0.01326291  0.46188895  3   3 -0.20842081  0.3662038
#> 10 -0.30420261  0.19521967  3   4 -0.01326291 -0.2084208
#> 11 -0.24722026 -0.18616474  4   1          NA         NA

dby(d,y ~ id + order(num), dlag, ARGS=list(k=1:2), NAMES=c("l1","l2"))

#>              y           x id num          l1         l2
#> 1  -1.79543538  0.55634005  1   1          NA         NA
#> 2   0.15513661  0.03833857  1   2 -1.79543538         NA
#> 3  -1.55336807 -2.03545050  1   3  0.15513661 -1.7954354
#> 4  -0.71603154 -0.06096876  2   1          NA         NA
#> 5  -0.87809866  1.05140078  2   2 -0.71603154         NA
#> 6   0.78476462 -0.11705507  2   3 -0.87809866 -0.7160315
#> 7   0.36620380 -1.21157397  3   1          NA         NA
#> 8  -0.20842081  1.16618593  3   2  0.36620380         NA
#> 9  -0.01326291  0.46188895  3   3 -0.20842081  0.3662038
#> 10 -0.30420261  0.19521967  3   4 -0.01326291 -0.2084208
#> 11 -0.24722026 -0.18616474  4   1          NA         NA

dby(d, y~id + order(num), mean=mean, csum=cumsum, n=length)

#>              y           x id num        mean       csum n
#> 1  -1.79543538  0.55634005  1   1 -1.06455561 -1.7954354 3
#> 2   0.15513661  0.03833857  1   2 -1.06455561 -1.6402988 3
#> 3  -1.55336807 -2.03545050  1   3 -1.06455561 -3.1936668 3
#> 4  -0.71603154 -0.06096876  2   1 -0.26978853 -0.7160315 3
#> 5  -0.87809866  1.05140078  2   2 -0.26978853 -1.5941302 3
#> 6   0.78476462 -0.11705507  2   3 -0.26978853 -0.8093656 3
#> 7   0.36620380 -1.21157397  3   1 -0.03992063  0.3662038 4
#> 8  -0.20842081  1.16618593  3   2 -0.03992063  0.1577830 4
#> 9  -0.01326291  0.46188895  3   3 -0.03992063  0.1445201 4
#> 10 -0.30420261  0.19521967  3   4 -0.03992063 -0.1596825 4
#> 11 -0.24722026 -0.18616474  4   1 -0.24722026 -0.2472203 1

dby(d2, y~id + order(num), a=cumsum, b=mean, N=length, l1=function(x) c(NA,x)[-length(x)])

#>              y           x id num          a           b N         l1
#> 1  -1.79543538  0.55634005  1   1 -1.7954354 -1.06455561 3         NA
#> 2   0.15513661  0.03833857  1   2 -1.6402988 -1.06455561 3 -1.7954354
#> 3  -1.55336807 -2.03545050  1   3 -3.1936668 -1.06455561 3 -1.5533681
#> 4  -0.71603154 -0.06096876  2   1 -0.7160315 -0.26978853 3         NA
#> 5  -0.87809866  1.05140078  2   2 -1.5941302 -0.26978853 3 -0.7160315
#> 6   0.78476462 -0.11705507  2   3 -0.8093656 -0.26978853 3  0.7847646
#> 7   0.36620380 -1.21157397  3   1  0.3662038 -0.03992063 4         NA
#> 8  -0.20842081  1.16618593  3   2  0.1577830 -0.03992063 4  0.3662038
#> 9  -0.01326291  0.46188895  3   3  0.1445201 -0.03992063 4 -0.2084208
#> 10 -0.30420261  0.19521967  3   4 -0.1596825 -0.03992063 4 -0.3042026
#> 11 -0.24722026 -0.18616474  4   1 -0.2472203 -0.24722026 1 -0.2472203

dby(d, y~id + order(num), nn=seq_along, n=length)

#>              y           x id num nn n
#> 1  -1.79543538  0.55634005  1   1  1 3
#> 2   0.15513661  0.03833857  1   2  2 3
#> 3  -1.55336807 -2.03545050  1   3  3 3
#> 4  -0.71603154 -0.06096876  2   1  1 3
#> 5  -0.87809866  1.05140078  2   2  2 3
#> 6   0.78476462 -0.11705507  2   3  3 3
#> 7   0.36620380 -1.21157397  3   1  1 4
#> 8  -0.20842081  1.16618593  3   2  2 4
#> 9  -0.01326291  0.46188895  3   3  3 4
#> 10 -0.30420261  0.19521967  3   4  4 4
#> 11 -0.24722026 -0.18616474  4   1  1 1

dby(d, y~id + order(num), nn=seq_along, n=length)

#>              y           x id num nn n
#> 1  -1.79543538  0.55634005  1   1  1 3
#> 2   0.15513661  0.03833857  1   2  2 3
#> 3  -1.55336807 -2.03545050  1   3  3 3
#> 4  -0.71603154 -0.06096876  2   1  1 3
#> 5  -0.87809866  1.05140078  2   2  2 3
#> 6   0.78476462 -0.11705507  2   3  3 3
#> 7   0.36620380 -1.21157397  3   1  1 4
#> 8  -0.20842081  1.16618593  3   2  2 4
#> 9  -0.01326291  0.46188895  3   3  3 4
#> 10 -0.30420261  0.19521967  3   4  4 4
#> 11 -0.24722026 -0.18616474  4   1  1 1

d <- d[,1:4]
dby(d, x<0) <- list(z=mean)

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

d <- dby(d, is.na(z), z=1)

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

f <- function(x) apply(x,1,min)
dby(d, y+x~id, min=f)

#> Error in .ApplyBy2(INPUT, ID, F = expr, Env = env, Argument = "x", Columnwise = COLUMN): object 'f' not found

dby(d,y+x~id+order(num), function(x) x)

#>              y           x id num         z         _11         _12
#> 1  -1.79543538  0.55634005  1   1 1.0000000 -1.79543538  0.55634005
#> 2   0.15513661  0.03833857  1   2 1.0000000  0.15513661  0.03833857
#> 3  -1.55336807 -2.03545050  1   3 0.8011568 -1.55336807 -2.03545050
#> 4  -0.71603154 -0.06096876  2   1 0.8011568 -0.71603154 -0.06096876
#> 5  -0.87809866  1.05140078  2   2 1.0000000 -0.87809866  1.05140078
#> 6   0.78476462 -0.11705507  2   3 0.8011568  0.78476462 -0.11705507
#> 7   0.36620380 -1.21157397  3   1 0.8011568  0.36620380 -1.21157397
#> 8  -0.20842081  1.16618593  3   2 1.0000000 -0.20842081  1.16618593
#> 9  -0.01326291  0.46188895  3   3 1.0000000 -0.01326291  0.46188895
#> 10 -0.30420261  0.19521967  3   4 1.0000000 -0.30420261  0.19521967
#> 11 -0.24722026 -0.18616474  4   1 0.8011568 -0.24722026 -0.18616474

f <- function(x) { cbind(cumsum(x[,1]),cumsum(x[,2]))/sum(x)}
dby(d, y+x~id, f)

#> Error in .ApplyBy2(INPUT, ID, F = expr, Env = env, Argument = "x", Columnwise = COLUMN): object 'f' not found

## column-wise
a <- d
dby2(a, mean, median, REGEX=TRUE) <- '^[y|x]'~id
a

#>              y           x id num         z      mean.y     mean.x   median.y
#> 1  -1.79543538  0.55634005  1   1 1.0000000 -1.06455561 -0.4802573 -1.5533681
#> 2   0.15513661  0.03833857  1   2 1.0000000 -1.06455561 -0.4802573 -1.5533681
#> 3  -1.55336807 -2.03545050  1   3 0.8011568 -1.06455561 -0.4802573 -1.5533681
#> 4  -0.71603154 -0.06096876  2   1 0.8011568 -0.26978853  0.2911257 -0.7160315
#> 5  -0.87809866  1.05140078  2   2 1.0000000 -0.26978853  0.2911257 -0.7160315
#> 6   0.78476462 -0.11705507  2   3 0.8011568 -0.26978853  0.2911257 -0.7160315
#> 7   0.36620380 -1.21157397  3   1 0.8011568 -0.03992063  0.1529301 -0.1108419
#> 8  -0.20842081  1.16618593  3   2 1.0000000 -0.03992063  0.1529301 -0.1108419
#> 9  -0.01326291  0.46188895  3   3 1.0000000 -0.03992063  0.1529301 -0.1108419
#> 10 -0.30420261  0.19521967  3   4 1.0000000 -0.03992063  0.1529301 -0.1108419
#> 11 -0.24722026 -0.18616474  4   1 0.8011568 -0.24722026 -0.1861647 -0.2472203
#>       median.x
#> 1   0.03833857
#> 2   0.03833857
#> 3   0.03833857
#> 4  -0.06096876
#> 5  -0.06096876
#> 6  -0.06096876
#> 7   0.32855431
#> 8   0.32855431
#> 9   0.32855431
#> 10  0.32855431
#> 11 -0.18616474

## wildcards 
dby2(a,'y*'+'x*'~id,mean)

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

#>              y           x id num         z   median.y    median.x      mean.y
#> 1  -1.79543538  0.55634005  1   1 1.0000000 -1.5533681  0.03833857 -1.06455561
#> 2   0.15513661  0.03833857  1   2 1.0000000 -1.5533681  0.03833857 -1.06455561
#> 3  -1.55336807 -2.03545050  1   3 0.8011568 -1.5533681  0.03833857 -1.06455561
#> 4  -0.71603154 -0.06096876  2   1 0.8011568 -0.7160315 -0.06096876 -0.26978853
#> 5  -0.87809866  1.05140078  2   2 1.0000000 -0.7160315 -0.06096876 -0.26978853
#> 6   0.78476462 -0.11705507  2   3 0.8011568 -0.7160315 -0.06096876 -0.26978853
#> 7   0.36620380 -1.21157397  3   1 0.8011568 -0.1108419  0.32855431 -0.03992063
#> 8  -0.20842081  1.16618593  3   2 1.0000000 -0.1108419  0.32855431 -0.03992063
#> 9  -0.01326291  0.46188895  3   3 1.0000000 -0.1108419  0.32855431 -0.03992063
#> 10 -0.30420261  0.19521967  3   4 1.0000000 -0.1108419  0.32855431 -0.03992063
#> 11 -0.24722026 -0.18616474  4   1 0.8011568 -0.2472203 -0.18616474 -0.24722026
#>        mean.x
#> 1  -0.4802573
#> 2  -0.4802573
#> 3  -0.4802573
#> 4   0.2911257
#> 5   0.2911257
#> 6   0.2911257
#> 7   0.1529301
#> 8   0.1529301
#> 9   0.1529301
#> 10  0.1529301
#> 11 -0.1861647

## subset
dby(d, x<0) <- list(z=NA)

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

d

#>              y           x id num  z
#> 1  -1.79543538  0.55634005  1   1  1
#> 2   0.15513661  0.03833857  1   2  1
#> 3  -1.55336807 -2.03545050  1   3 NA
#> 4  -0.71603154 -0.06096876  2   1 NA
#> 5  -0.87809866  1.05140078  2   2  1
#> 6   0.78476462 -0.11705507  2   3 NA
#> 7   0.36620380 -1.21157397  3   1 NA
#> 8  -0.20842081  1.16618593  3   2  1
#> 9  -0.01326291  0.46188895  3   3  1
#> 10 -0.30420261  0.19521967  3   4  1
#> 11 -0.24722026 -0.18616474  4   1 NA

dby(d, y~id|x>-1, v=mean,z=1)

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

#>              y           x id num          v  z
#> 1  -1.79543538  0.55634005  1   1 -0.8201494  1
#> 2   0.15513661  0.03833857  1   2 -0.8201494  1
#> 3  -1.55336807 -2.03545050  1   3         NA NA
#> 4  -0.71603154 -0.06096876  2   1 -0.2697885  1
#> 5  -0.87809866  1.05140078  2   2 -0.2697885  1
#> 6   0.78476462 -0.11705507  2   3 -0.2697885  1
#> 7   0.36620380 -1.21157397  3   1         NA NA
#> 8  -0.20842081  1.16618593  3   2 -0.1752954  1
#> 9  -0.01326291  0.46188895  3   3 -0.1752954  1
#> 10 -0.30420261  0.19521967  3   4 -0.1752954  1
#> 11 -0.24722026 -0.18616474  4   1 -0.2472203  1

dby(d, y+x~id|x>-1, mean, median, COLUMN=TRUE)

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

#>              y           x id num  z     mean.y     mean.x   median.y
#> 1  -1.79543538  0.55634005  1   1  1 -0.8201494  0.2973393 -0.8201494
#> 2   0.15513661  0.03833857  1   2  1 -0.8201494  0.2973393 -0.8201494
#> 3  -1.55336807 -2.03545050  1   3 NA         NA         NA         NA
#> 4  -0.71603154 -0.06096876  2   1 NA -0.2697885  0.2911257 -0.7160315
#> 5  -0.87809866  1.05140078  2   2  1 -0.2697885  0.2911257 -0.7160315
#> 6   0.78476462 -0.11705507  2   3 NA -0.2697885  0.2911257 -0.7160315
#> 7   0.36620380 -1.21157397  3   1 NA         NA         NA         NA
#> 8  -0.20842081  1.16618593  3   2  1 -0.1752954  0.6077648 -0.2084208
#> 9  -0.01326291  0.46188895  3   3  1 -0.1752954  0.6077648 -0.2084208
#> 10 -0.30420261  0.19521967  3   4  1 -0.1752954  0.6077648 -0.2084208
#> 11 -0.24722026 -0.18616474  4   1 NA -0.2472203 -0.1861647 -0.2472203
#>       median.x
#> 1   0.29733931
#> 2   0.29733931
#> 3           NA
#> 4  -0.06096876
#> 5  -0.06096876
#> 6  -0.06096876
#> 7           NA
#> 8   0.46188895
#> 9   0.46188895
#> 10  0.46188895
#> 11 -0.18616474

dby2(d, y+x~id|x>0, mean, REDUCE=TRUE)

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

#>   id     mean.y    mean.x
#> 1  1 -0.8201494 0.2973393
#> 2  2 -0.8780987 1.0514008
#> 3  3 -0.1752954 0.6077648

dby(d,y~id|x<0,mean,ALL=FALSE)

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.

#>             y           x id num  z        mean
#> 3  -1.5533681 -2.03545050  1   3 NA -1.55336807
#> 4  -0.7160315 -0.06096876  2   1 NA  0.03436654
#> 6   0.7847646 -0.11705507  2   3 NA  0.03436654
#> 7   0.3662038 -1.21157397  3   1 NA  0.36620380
#> 11 -0.2472203 -0.18616474  4   1 NA -0.24722026

a <- iris
a <- dby(a,y=1)
dby(a,Species=="versicolor") <- list(y=2)

#> Warning: Using formula(x) is deprecated when x is a character vector of length > 1.
#>   Consider formula(paste(x, collapse = " ")) instead.