Split a data set and run function from timereg and aggregate

Split a data set and run function of cox-aalen type and aggregate results

divide.conquer.timereg(func = NULL, data, size, ...)

Arguments

func

called function

data

data-frame

size

size of splits

...

Additional arguments to lower level functions

Author

Thomas Scheike, Klaus K. Holst

Examples

library(timereg)
data(TRACE)
a <- divide.conquer.timereg(prop.odds,TRACE,
                            formula=Event(time,status==9)~chf+vf+age,n.sim=0,size=200)
coef(a)
#>      Coef.      SE Robust SE D2log(L)^-1    z   P-val lower2.5% upper97.5%
#> chf 1.1100 0.09990    0.0999           1 11.1 0.0e+00    0.9140     1.3100
#> vf  1.3300 0.20100    0.2250           1  5.9 3.7e-09    0.9360     1.7200
#> age 0.0726 0.00527    0.0052           1 14.0 0.0e+00    0.0623     0.0829
a2 <- divide.conquer.timereg(prop.odds,TRACE,
                             formula=Event(time,status==9)~chf+vf+age,n.sim=0,size=500)
coef(a2)
#>      Coef.      SE Robust SE D2log(L)^-1     z    P-val lower2.5% upper97.5%
#> chf 1.0900 0.09900   0.09910           1 11.00 0.00e+00    0.8960     1.2800
#> vf  1.3900 0.19100   0.21300           1  6.55 5.88e-11    1.0200     1.7600
#> age 0.0727 0.00534   0.00527           1 13.80 0.00e+00    0.0622     0.0832

if (interactive()) {
par(mfrow=c(1,1))
plot(a,xlim=c(0,8),ylim=c(0,0.01))
par(new=TRUE)
plot(a2,xlim=c(0,8),ylim=c(0,0.01))
}