list, head, print, tail

listing for data frames

Usage

dprint(data, y = NULL, n = 0, ..., x = NULL)

Arguments

data

if x is formula or names for data frame then data frame is needed.

y

name of variable, or fomula, or names of variables on data frame.

n

Index of observations to print (default c(1:nfirst, n-nlast:nlast)

...

Optional additional arguments (nfirst,nlast, and print options)

x

possible group variable

Author

Klaus K. Holst and Thomas Scheike

Examples

m <- lava::lvm(letters)
d <- lava::sim(m, 20)

dlist(d,~a+b+c)
#>     a       b       c      
#> 1   -0.3430  0.6540 -0.2336
#> 2    0.0318  0.2201  0.4094
#> 3   -0.6842  0.2222 -0.2991
#> 4    1.5426 -0.7988 -0.6528
#> 5    0.9218 -0.7783  1.4532
#> ---                        
#> 16  -1.6899  1.0053 -0.3569
#> 17  -1.2498 -0.5929  0.7844
#> 18   0.2165 -0.7282  1.3261
#> 19   0.2049 -0.3988  0.4502
#> 20  -1.2911 -0.8080 -1.4074
dlist(d,~a+b+c|a<0 & b>0)
#>    a       b       c      
#> 1  -0.3430 0.65400 -0.2336
#> 3  -0.6842 0.22218 -0.2991
#> 6  -1.1799 0.20327 -0.7293
#> 7  -0.2051 0.03845  1.5550
#> 15 -2.0719 0.11478  1.6434
#> 16 -1.6899 1.00526 -0.3569
## listing all : 
dlist(d,~a+b+c|a<0 & b>0,n=0)
#>             a         b          c
#> 1  -0.3429606 0.6540016 -0.2336226
#> 3  -0.6842156 0.2221841 -0.2990729
#> 6  -1.1798613 0.2032687 -0.7293024
#> 7  -0.2051132 0.0384453  1.5550198
#> 15 -2.0718854 0.1147781  1.6433526
#> 16 -1.6898776 1.0052562 -0.3569498
dlist(d,a+b+c~I(d>0)|a<0 & b>0)
#> I(d > 0): FALSE
#>   a       b       c      
#> 1 -0.3430 0.65400 -0.2336
#> 7 -0.2051 0.03845  1.5550
#> ------------------------------------------------------------ 
#> I(d > 0): TRUE
#>    a       b      c      
#> 3  -0.6842 0.2222 -0.2991
#> 6  -1.1799 0.2033 -0.7293
#> 15 -2.0719 0.1148  1.6434
#> 16 -1.6899 1.0053 -0.3569
dlist(d,.~I(d>0)|a<0 & b>0)
#> I(d > 0): FALSE
#>   a       b       c       d      e       f       g       h     i       j      
#> 1 -0.3430 0.65400 -0.2336 -1.140  0.2367  1.2253 -0.5191 1.776 -0.8307  0.2164
#> 7 -0.2051 0.03845  1.5550 -1.037 -1.7861 -0.4491 -0.6595 1.735  0.6057 -0.7904
#>   k       l       m       n       o      p     q       r       s       t     
#> 1 -3.0725  3.0338 -1.4461 -1.8496 -1.066 1.094 -0.2594 -0.1680 -0.1453 0.6335
#> 7  0.8285 -0.6438 -0.6378 -0.7651 -2.502 1.465  1.0925  0.6029 -0.8935 0.5941
#>   u       v      w       x       y      z      
#> 1 -0.8547 -1.740  0.9347  0.1864 0.3096 -0.2575
#> 7 -1.8431 -0.403 -0.5314 -0.7250 0.7838 -0.4264
#> ------------------------------------------------------------ 
#> I(d > 0): TRUE
#>    a       b      c       d      e          f       g       h       i       
#> 3  -0.6842 0.2222 -0.2991 0.3541  1.1628378  0.5754 -0.3504 -0.4445 -0.01373
#> 6  -1.1799 0.2033 -0.7293 1.5257  0.0002668 -1.1034 -0.9070 -1.0868 -1.56290
#> 15 -2.0719 0.1148  1.6434 0.6863  0.4499602 -0.9480 -1.4678 -1.2130  0.60040
#> 16 -1.6899 1.0053 -0.3569 1.9157 -0.6830655 -1.3723 -0.4467 -0.8346  0.93348
#>    j       k       l       m       n        o       p       q       r      
#> 3  -0.5947  0.8106 -0.8047  0.2040 -0.52619 -0.5382 -1.3569  1.1293  0.3438
#> 6  -0.9040  1.1354 -0.3637  1.9247  0.33472  2.0164  0.5303  0.5090 -1.8195
#> 15  1.6412 -1.5693 -1.4313  0.3385  0.03272  1.4015 -1.0211 -1.6163  1.7320
#> 16  0.2203  0.8887 -0.3970 -0.3253  1.52795  1.1233 -0.3984 -0.7271  0.5447
#>    s       t       u        v      w        x       y       z     
#> 3  -0.1836 -0.2914 -0.08432  0.306 -0.60139 -0.4747 -2.7242  1.531
#> 6  -0.7420 -1.4042 -0.56224  1.090  0.02123 -1.6962 -0.1941  0.227
#> 15  0.5249  0.3813  1.17399  1.430  0.08390  0.1962  0.3495 -1.495
#> 16 -0.6410  0.9947  0.34113 -1.046  0.26112 -0.7878  0.8414 -1.814
dlist(d,~a+b+c|a<0 & b>0, nlast=0)
#>    a       b       c      
#> 1  -0.3430 0.65400 -0.2336
#> 3  -0.6842 0.22218 -0.2991
#> 6  -1.1799 0.20327 -0.7293
#> 7  -0.2051 0.03845  1.5550
#> 15 -2.0719 0.11478  1.6434
dlist(d,~a+b+c|a<0 & b>0, nfirst=3, nlast=3)
#>    a       b       c      
#> 1  -0.3430 0.65400 -0.2336
#> 3  -0.6842 0.22218 -0.2991
#> 6  -1.1799 0.20327 -0.7293
#> 7  -0.2051 0.03845  1.5550
#> 15 -2.0719 0.11478  1.6434
#> 16 -1.6899 1.00526 -0.3569
dlist(d,~a+b+c|a<0 & b>0, 1:5)
#>    a       b       c      
#> 1  -0.3430 0.65400 -0.2336
#> 3  -0.6842 0.22218 -0.2991
#> 6  -1.1799 0.20327 -0.7293
#> 7  -0.2051 0.03845  1.5550
#> 15 -2.0719 0.11478  1.6434
dlist(d,~a+b+c|a<0 & b>0, -(5:1))
#>    a       b       c      
#> 3  -0.6842 0.22218 -0.2991
#> 6  -1.1799 0.20327 -0.7293
#> 7  -0.2051 0.03845  1.5550
#> 15 -2.0719 0.11478  1.6434
#> 16 -1.6899 1.00526 -0.3569
dlist(d,~a+b+c|a<0 & b>0, list(1:5,50:55,-(5:1)))
#>     a       b       c      
#> 1   -0.3430 0.65400 -0.2336
#> 3   -0.6842 0.22218 -0.2991
#> 6   -1.1799 0.20327 -0.7293
#> 7   -0.2051 0.03845  1.5550
#> 15  -2.0719 0.11478  1.6434
#> ---                        
#> ---                        
#> 3   -0.6842 0.22218 -0.2991
#> 6   -1.1799 0.20327 -0.7293
#> 7   -0.2051 0.03845  1.5550
#> 15  -2.0719 0.11478  1.6434
#> 16  -1.6899 1.00526 -0.3569
dprint(d,a+b+c ~ I(d>0) |a<0 & b>0, list(1:5,50:55,-(5:1)))
#> I(d > 0): FALSE
#>     a       b       c      
#> 1   -0.3430 0.65400 -0.2336
#> 7   -0.2051 0.03845  1.5550
#> ---                        
#> ---                        
#> 1   -0.3430 0.65400 -0.2336
#> 7   -0.2051 0.03845  1.5550
#> ------------------------------------------------------------ 
#> I(d > 0): TRUE
#>     a       b      c      
#> 3   -0.6842 0.2222 -0.2991
#> 6   -1.1799 0.2033 -0.7293
#> 15  -2.0719 0.1148  1.6434
#> 16  -1.6899 1.0053 -0.3569
#> ---                       
#> ---                       
#> 3   -0.6842 0.2222 -0.2991
#> 6   -1.1799 0.2033 -0.7293
#> 15  -2.0719 0.1148  1.6434
#> 16  -1.6899 1.0053 -0.3569