Wrapper for easy fitting of Clayton-Oakes or bivariate Plackett models for bivariate survival data

Fits two-stage model for describing depdendence in survival data using marginals that are on cox or aalen form using the twostage funcion, but call is different and easier and the data manipulation build into the function. Useful in particular for family design data.

easy.survival.twostage(
  margsurv = NULL,
  data = parent.frame(),
  method = "nlminb",
  status = "status",
  time = "time",
  entry = NULL,
  id = "id",
  Nit = 60,
  detail = 0,
  silent = 1,
  weights = NULL,
  control = list(),
  theta = NULL,
  theta.formula = NULL,
  desnames = NULL,
  deshelp = 0,
  var.link = 1,
  step = 0.5,
  model = "plackett",
  marginal.surv = NULL,
  strata = NULL,
  se.clusters = NULL
)

Arguments

margsurv

model

data

data frame

method

Scoring method

status

Status at exit time

time

Exit time

entry

Entry time

id

name of cluster variable in data frame

Nit

Number of iterations

detail

Detail for more output for iterations

silent

Debug information

weights

Weights for log-likelihood, can be used for each type of outcome in 2x2 tables.

control

Optimization arguments

theta

Starting values for variance components

theta.formula

design for depedence, either formula or design function

desnames

names for dependence parameters

deshelp

if 1 then prints out some data sets that are used, on on which the design function operates

var.link

Link function for variance (exp link)

step

Step size for newton-raphson

model

plackett or clayton-oakes model

marginal.surv

vector of marginal survival probabilities

strata

strata for fitting

se.clusters

clusters for iid decomposition for roubst standard errors

Details

If clusters contain more than two times, the algoritm uses a composite likelihood based on the pairwise bivariate models.

The reported standard errors are based on the estimated information from the likelihood assuming that the marginals are known.

Examples

library(mets)
data("prt",package="mets")
prtsam <- blocksample(prt,idvar="id",1e3,replace=FALSE)
margp <- coxph(Surv(time,status==1)~factor(country),data=prtsam)
fitco <- survival.twostage(margp,data=prtsam,clusters=prtsam$id)
summary(fitco)
#> Dependence parameter for Clayton-Oakes model
#> Variance of Gamma distributed random effects 
#> With log-link 
#> $estimates
#>              log-Coef.        SE         z       P-val Kendall tau         SE
#> dependence1 -0.9363086 0.3038615 -3.081366 0.002060531   0.1639049 0.04164121
#> 
#> $vargam
#>             Estimate Std.Err   2.5%  97.5%   P-value
#> dependence1   0.3921  0.1191 0.1586 0.6256 0.0009984
#> 
#> $type
#> [1] "clayton.oakes"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

des <- model.matrix(~-1+factor(zyg),data=prtsam);
fitco <- survival.twostage(margp,data=prtsam,theta.des=des,clusters=prtsam$id)
summary(fitco)
#> Dependence parameter for Clayton-Oakes model
#> Variance of Gamma distributed random effects 
#> With log-link 
#> $estimates
#>                log-Coef.        SE        z      P-val Kendall tau         SE
#> factor(zyg)DZ -0.9452397 0.3736229 -2.52993 0.01140853   0.1626847 0.05089430
#> factor(zyg)MZ -0.9245538 0.5038084 -1.83513 0.06648646   0.1655222 0.06958833
#> 
#> $vargam
#>             Estimate Std.Err    2.5%  97.5% P-value
#> dependence1   0.3886  0.1452 0.10403 0.6731 0.00744
#> dependence2   0.3967  0.1999 0.00498 0.7884 0.04716
#> 
#> $type
#> [1] "clayton.oakes"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"
rm(prtsam)

dfam <- simSurvFam(1000)
dfam <- fast.reshape(dfam,var=c("x","time","status"))

desfs <- function(x,num1="num1",num2="num2")
{
pp <- (x[num1]=="m")*(x[num2]=="f")*1   ## mother-father
pc <- (x[num1]=="m" | x[num1]=="f")*(x[num2]=="b1" | x[num2]=="b2")*1 ## mother-child
cc <- (x[num1]=="b1")*(x[num2]=="b1" | x[num2]=="b2")*1               ## child-child
c(pp,pc,cc)
}

marg <- coxph(Surv(time,status)~factor(num),data=dfam)
out3 <- easy.survival.twostage(marg,data=dfam,time="time",status="status",id="id",
             deshelp=0,
             method="nr",theta.formula=desfs,
             model="plackett",
             desnames=c("parent-parent","parent-child","child-cild"))
summary(out3)
#> Dependence parameter for Odds-Ratio (Plackett) model 
#> With log-link 
#> $estimates
#>               log-Coef.         SE        z P-val Spearman Corr.         SE
#> parent-parent 1.6674544 0.08834856 18.87359     0      0.5089844 0.02245865
#> parent-child  0.7751126 0.06336616 12.23228     0      0.2533054 0.01989718
#> child-cild    0.9293058 0.08872420 10.47410     0      0.3011186 0.02714681
#> 
#> $or
#>               Estimate Std.Err  2.5% 97.5%   P-value
#> parent-parent    5.299  0.4681 4.381 6.216 1.059e-29
#> parent-child     2.171  0.1376 1.901 2.440 4.185e-56
#> child-cild       2.533  0.2247 2.092 2.973 1.827e-29
#> 
#> $type
#> [1] "plackett"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"