mets: Analysis of Multivariate Event Times

Implementation of various statistical models for multivariate event history data doi:10.1007/s10985-013-9244-x . Including multivariate cumulative incidence models doi:10.1002/sim.6016 , and bivariate random effects probit models (Liability models) doi:10.1016/j.csda.2015.01.014 . Modern methods for survival analysis, including regression modelling (Cox, Fine-Gray, Ghosh-Lin, Binomial regression) with fast computation of influence functions.

Implementation of various statistical models for multivariate event history data. Including multivariate cumulative incidence models, and bivariate random effects probit models (Liability models)

See also

Author

Maintainer: Klaus K. Holst klaus@holst.it

Authors:

Thomas Scheike

Klaus K. Holst and Thomas Scheike

Examples

## To appear