Lu-Tsiatis More Efficient Log-Rank for Randomized studies with baseline covariates

Efficient implementation of the Lu-Tsiatis improvement using baseline covariates. Results almost equivalent with the speffSurv function of the speff2trial function. A dynamic censoring augmentation regression is also computed to gain even more from the censoring augmentation.

phreg_lt(
  formula,
  data,
  augmentR = NULL,
  treat.model = ~+1,
  augmentC = NULL,
  km = TRUE,
  cens.code = 0,
  level = 0.95,
  cens.model = NULL,
  typeII = NULL,
  ...
)

Arguments

formula

formula with 'Surv' outcome (see coxph) and treatment (randomization 0/1)

data

data frame

augmentR

formula for the randomization augmentation (~age+sex)

treat.model

propensity score model, default is ~+1, assuming RCT study

augmentC

formula for the censoring augmentation (~age+sex)

km

use Kaplan-Meier for the censoring weights (stratified on treatment)

cens.code

censoring code

level

of confidence intervals

cens.model,

default is censoring model ~strata(treatment) but any model can be used to make censoring martingales

typeII

if 1 then computes also alternative formulae that are based on the censoring martingale rather than the robust processes of Lu-Tsiatis computations.

...

Additional arguments to phreg function

References

Lu, Tsiatis (2008), Improving the efficiency of the log-rank test using auxiliary covariates, Biometrika, 679--694

Author

Thomas Scheike

Examples

## Lu, Tsiatis simulation
data <- mets:::simLT(0.7,100)

out <- phreg_lt(Surv(time,status)~Z,data=data,augmentR=~X,augmentC=~factor(Z):X)
out$coefs
#>               Estimate   Std.Err       2.5%     97.5%   P-value Var-Ratio
#> Marginal-Z   0.1722827 0.2727585 -0.3623143 0.7068796 0.5276282 1.0000000
#> Lu-Tsiatis-Z 0.1060313 0.2291500 -0.3430944 0.5551569 0.6435683 0.7058020
#> LT-AugR-Z    0.1452953 0.2301751 -0.3058396 0.5964302 0.5278848 0.7121312
#> LT-AugC-Z    0.1331333 0.2710266 -0.3980691 0.6643357 0.6232719 0.9873410
#> Dynamic-LT-Z 0.0236330 0.2194682 -0.4065168 0.4537828 0.9142471 0.6474205