Augmentation for Binomial regression based on stratified NPMLE Cif (Aalen-Johansen)

Computes the augmentation term for each individual as well as the sum $$ A = \int_0^t H(u,X) \frac{1}{S^*(u,s)} \frac{1}{G_c(u)} dM_c(u) $$ with $$ H(u,X) = F_1^*(t,s) - F_1^*(u,s) $$ using a KM for $$G_c(t)$$ and a working model for cumulative baseline related to $$F_1^*(t,s)$$ and $$s$$ is strata, $$S^*(t,s) = 1 - F_1^*(t,s) - F_2^*(t,s)$$.

BinAugmentCifstrata(
  formula,
  data = data,
  cause = 1,
  cens.code = 0,
  km = TRUE,
  time = NULL,
  weights = NULL,
  offset = NULL,
  ...
)

Arguments

formula
formula with 'Event', strata model for CIF given by strata, and strataC specifies censoring strata

data
data frame

cause
of interest

cens.code
code of censoring

km
to use Kaplan-Meier

time
of interest

weights
weights for estimating equations

offset
offsets for logistic regression

...
Additional arguments to binreg function.

Details

Standard errors computed under assumption of correct $$G_c(s)$$ model.

Author

Thomas Scheike

Examples

data(bmt)
dcut(bmt,breaks=2) <- ~age 
out1<-BinAugmentCifstrata(Event(time,cause)~platelet+agecat.2+
      strata(platelet,agecat.2),data=bmt,cause=1,time=40)
summary(out1)

#> 
#>    n events
#>  408    157
#> 
#>  408 clusters
#> coeffients:
#>                      Estimate  Std.Err     2.5%    97.5% P-value
#> (Intercept)          -0.50044  0.17069 -0.83498 -0.16590  0.0034
#> platelet             -0.63447  0.23596 -1.09695 -0.17200  0.0072
#> agecat.2(0.203,1.94]  0.53823  0.21192  0.12287  0.95358  0.0111
#> 
#> exp(coeffients):
#>                      Estimate    2.5%  97.5%
#> (Intercept)           0.60626 0.43388 0.8471
#> platelet              0.53021 0.33389 0.8420
#> agecat.2(0.203,1.94]  1.71297 1.13074 2.5950
#> 
#> 

out2<-BinAugmentCifstrata(Event(time,cause)~platelet+agecat.2+
    strata(platelet,agecat.2)+strataC(platelet),data=bmt,cause=1,time=40)

#> Warning: NaNs produced

summary(out2)

#> 
#>    n events
#>  408    157
#> 
#>  408 clusters
#> coeffients:
#>                      Estimate  Std.Err     2.5%    97.5% P-value
#> (Intercept)          -0.50015  0.17118 -0.83566 -0.16463  0.0035
#> platelet             -0.63540  0.23741 -1.10071 -0.17010  0.0074
#> agecat.2(0.203,1.94]  0.53768  0.21227  0.12163  0.95372  0.0113
#> 
#> exp(coeffients):
#>                      Estimate    2.5%  97.5%
#> (Intercept)           0.60644 0.43359 0.8482
#> platelet              0.52972 0.33264 0.8436
#> agecat.2(0.203,1.94]  1.71203 1.12934 2.5953
#> 
#>