Augmentation for Binomial regression based on stratified NPMLE Cif (Aalen-Johansen)

Computes the augmentation term for each individual as well as the sum $$ A = \int_0^t H(u,X) \frac{1}{S^*(u,s)} \frac{1}{G_c(u)} dM_c(u) $$ with $$ H(u,X) = F_1^*(t,s) - F_1^*(u,s) $$ using a KM for $$G_c(t)$$ and a working model for cumulative baseline related to $$F_1^*(t,s)$$ and $$s$$ is strata, $$S^*(t,s) = 1 - F_1^*(t,s) - F_2^*(t,s)$$.

BinAugmentCifstrata(
  formula,
  data = data,
  cause = 1,
  cens.code = 0,
  km = TRUE,
  time = NULL,
  weights = NULL,
  offset = NULL,
  ...
)

Arguments

formula
formula with 'Event', strata model for CIF given by strata, and strataC specifies censoring strata

data
data frame

cause
of interest

cens.code
code of censoring

km
to use Kaplan-Meier

time
of interest

weights
weights for estimating equations

offset
offsets for logistic regression

...
Additional arguments to binreg function.

Details

Standard errors computed under assumption of correct $$G_c(s)$$ model.

Augmentation term only computed for standard FG model, since strata is used to specify working models for CIF's.

Examples

data(bmt)
dcut(bmt,breaks=2) <- ~age
out1<-BinAugmentCifstrata(Event(time,cause)~platelet+agecat.2+
                  strata(platelet,agecat.2),data=bmt,cause=1,time=40)
summary(out1)

#> 
#>    n events
#>  408    157
#> 
#>  408 clusters
#> log-coeffients:
#>                      Estimate  Std.Err     2.5%    97.5% P-value
#> (Intercept)          -0.50060  0.17070 -0.83517 -0.16603  0.0034
#> platelet             -0.63520  0.23597 -1.09769 -0.17271  0.0071
#> agecat.2(0.203,1.94]  0.53808  0.21193  0.12271  0.95344  0.0111
#> 
#> exp(coeffients):
#>                        Estimate Std.Err    2.5%   97.5% P-value
#> [(Intercept)]           0.60617 0.10347 0.40336 0.80897  0.0001
#> [platelet]              0.52983 0.12502 0.28479 0.77487  0.0002
#> [agecat.2(0.203,1.94]]  1.71271 0.36297 1.00131 2.42411  0.0496
#> 

out2<-BinAugmentCifstrata(Event(time,cause)~platelet+agecat.2+
    strata(platelet,agecat.2)+strataC(platelet),data=bmt,cause=1,time=40)

#> Warning: NaNs produced

summary(out2)

#> 
#>    n events
#>  408    157
#> 
#>  408 clusters
#> log-coeffients:
#>                      Estimate  Std.Err     2.5%    97.5% P-value
#> (Intercept)          -0.50014  0.17121 -0.83570 -0.16457  0.0035
#> platelet             -0.63664  0.23730 -1.10173 -0.17155  0.0073
#> agecat.2(0.203,1.94]  0.53724  0.21225  0.12123  0.95325  0.0114
#> 
#> exp(coeffients):
#>                        Estimate Std.Err    2.5%   97.5% P-value
#> [(Intercept)]           0.60645 0.10383 0.40295 0.80995  0.0002
#> [platelet]              0.52907 0.12555 0.28300 0.77513  0.0002
#> [agecat.2(0.203,1.94]]  1.71128 0.36323 0.99937 2.42319  0.0502
#>