Augmentation for Fine-Gray model based on stratified NPMLE Cif (Aalen-Johansen)

Computes the augmentation term for each individual as well as the sum $$ A(\beta) = \int H(t,X,\beta) \frac{F_2^*(t,s)}{S^*(t,s)} \frac{1}{G_c(t)} dM_c $$ with $$ H(t,X,\beta) = \int_t^\infty (X - E(\beta,t) ) G_c(t) d\Lambda_1^*i(t,s) $$ using a KM for $$G_c(t)$$ and a working model for cumulative baseline related to $$F_1^*(t,s)$$ and $$s$$ is strata, $$S^*(t,s) = 1 - F_1^*(t,s) - F_2^*(t,s)$$, and $$E(\beta^p,t)$$ is given.

FG_AugmentCifstrata(
  formula,
  data = data,
  E = NULL,
  cause = NULL,
  cens.code = 0,
  km = TRUE,
  case.weights = NULL,
  weights = NULL,
  offset = NULL,
  ...
)

Arguments

formula
formula with 'Event', strata model for CIF given by strata, and strataC specifies censoring strata

data
data frame

E
from FG-model

cause
of interest

cens.code
code of censoring

km
to use Kaplan-Meier

case.weights
weights for FG score equations (that follow dN_1)

weights
weights for FG score equations

offset
offsets for FG model

...
Additional arguments to lower level funtions

Details

After a couple of iterations we end up with a solution of $$ \int (X - E(\beta) ) Y_1(t) w(t) dM_1 + A(\beta) $$ the augmented FG-score.

Standard errors computed under assumption of correct $$G_c$$ model.

Examples

set.seed(100)
rho1 <- 0.2; rho2 <- 10
n <- 400
beta=c(0.0,-0.1,-0.5,0.3)
dats <- simul.cifs(n,rho1,rho2,beta,rc=0.2)
dtable(dats,~status)

#> 
#> status
#>   0   1   2 
#>  14  54 332 
#> 

dsort(dats) <- ~time
fg <- cifreg(Event(time,status)~Z1+Z2,data=dats,cause=1,propodds=NULL)
summary(fg)

#> 
#>    n events
#>  400     54
#> 
#>  400 clusters
#> log-coeffients:
#>     Estimate      S.E.   dU^-1/2 P-value
#> Z1  0.028262  0.135312  0.136188  0.8346
#> Z2 -0.149224  0.271222  0.272379  0.5822
#> 
#> exp(coeffients):
#>      Estimate Std.Err    2.5%   97.5% P-value
#> [Z1]  1.02866 0.13919 0.75586 1.30147  0.8368
#> [Z2]  0.86138 0.23362 0.40348 1.31927  0.5529
#> 

fgaugS <- FG_AugmentCifstrata(Event(time,status)~Z1+Z2+strata(Z1,Z2),data=dats,cause=1,E=fg$E)
summary(fgaugS)

#> 
#>    n events
#>  400     54
#> 
#>  400 clusters
#> log-coeffients:
#>      Estimate       S.E.    dU^-1/2 P-value
#> Z1  0.0088234  0.1294335  0.1361155  0.9457
#> Z2 -0.1738500  0.2600474  0.2725267  0.5038
#> 
#> exp(coeffients):
#>      Estimate Std.Err    2.5%   97.5% P-value
#> [Z1]  1.00886 0.13058 0.75293 1.26480  0.9459
#> [Z2]  0.84042 0.21855 0.41207 1.26877  0.4653
#> 

fgaugS2 <- FG_AugmentCifstrata(Event(time,status)~Z1+Z2+strata(Z1,Z2),data=dats,cause=1,E=fgaugS$E)
summary(fgaugS2)

#> 
#>    n events
#>  400     54
#> 
#>  400 clusters
#> log-coeffients:
#>      Estimate       S.E.    dU^-1/2 P-value
#> Z1  0.0081681  0.1294183  0.1361139  0.9497
#> Z2 -0.1746999  0.2600745  0.2725326  0.5018
#> 
#> exp(coeffients):
#>      Estimate Std.Err    2.5%   97.5% P-value
#> [Z1]  1.00820 0.13048 0.75247 1.26394  0.9499
#> [Z2]  0.83971 0.21839 0.41168 1.26774  0.4630
#>