Fast additive hazards model with robust standard errors

Fast Lin-Ying additive hazards model with a possibly stratified baseline. Robust variance is default variance with the summary.

aalenMets(formula, data = data, ...)

Arguments

formula

formula with 'Surv' outcome (see coxph)

data

data frame

...

Additional arguments to phreg

Details

influence functions (iid) will follow numerical order of given cluster variable so ordering after $id will give iid in order of data-set.

Author

Thomas Scheike

Examples

data(bmt)
out <- aalenMets(Surv(time,cause==1)~tcell+platelet+age,data=bmt)
summary(out)
#> 
#>    n events
#>  408    161
#> 
#>  408 clusters
#> coeffients:
#>            Estimate       S.E.    dU^-1/2 P-value
#> tcell    -0.0129551  0.0041287  0.2303841  0.0017
#> platelet -0.0087536  0.0028051  0.1664186  0.0018
#> age       0.0066176  0.0013879  0.0789182  0.0000
#> 
#> exp(coeffients):
#>          Estimate    2.5%  97.5%
#> tcell     0.98713 0.97917 0.9951
#> platelet  0.99128 0.98585 0.9967
#> age       1.00664 1.00390 1.0094
#> 
#>