Average Treatment effect for censored competing risks data using Binomial Regression

Under the standard causal assumptions we can estimate the average treatment effect E(Y(1) - Y(0)). We need Consistency, ignorability ( Y(1), Y(0) indep A given X), and positivity.

binregATE(
  formula,
  data,
  cause = 1,
  time = NULL,
  beta = NULL,
  treat.model = ~+1,
  cens.model = ~+1,
  offset = NULL,
  weights = NULL,
  cens.weights = NULL,
  se = TRUE,
  kaplan.meier = TRUE,
  cens.code = 0,
  no.opt = FALSE,
  method = "nr",
  augmentation = NULL,
  ...
)

Arguments

formula
formula with outcome (see coxph)

data
data frame

cause
cause of interest

time
time of interest

beta
starting values

treat.model
logistic treatment model given covariates

cens.model
only stratified cox model without covariates

offset
offsets for partial likelihood

weights
for score equations

cens.weights
censoring weights

se
to compute se's with IPCW adjustment, otherwise assumes that IPCW weights are known

kaplan.meier
uses Kaplan-Meier for IPCW in contrast to exp(-Baseline)

cens.code
gives censoring code

no.opt
to not optimize

method
for optimization

augmentation
to augment binomial regression

...
Additional arguments to lower level funtions

Details

The first covariate in the specification of the competing risks regression model must be the treatment effect that is binary. This is then model using a logistic regresssion using the standard binary double robust estimating equations that are then IPCW censoring adjusted using binomial regression.

Also computes the ATT and ATC, average treatment effect on the treated (ATT), E(Y(1) - Y(0) | A=1), and non-treated, respectively.

Rather than binomial regression we also consider a IPCW weighted version of standard logistic regression logitIPCWATE.

Author

Thomas Scheike

Examples

data(bmt)

brs <- binregATE(Event(time,cause)~tcell+platelet+age,bmt,time=50,cause=1,
  treat.model=tcell~platelet+age)
summary(brs)

#> 
#>    n events
#>  408    160
#> 
#>  408 clusters
#> coeffients:
#>              Estimate   Std.Err      2.5%     97.5% P-value
#> (Intercept) -0.195296  0.131037 -0.452123  0.061532  0.1361
#> tcell       -0.637346  0.359010 -1.340993  0.066301  0.0759
#> platelet    -0.351900  0.247162 -0.836329  0.132530  0.1545
#> age          0.419785  0.106904  0.210258  0.629312  0.0001
#> 
#> exp(coeffients):
#>             Estimate    2.5%  97.5%
#> (Intercept)  0.82259 0.63628 1.0635
#> tcell        0.52869 0.26159 1.0685
#> platelet     0.70335 0.43330 1.1417
#> age          1.52163 1.23400 1.8763
#> 
#> Average Treatment effects (G-formula) :
#>               Estimate    Std.Err       2.5%      97.5% P-value
#> treat-1      0.2896149  0.0664282  0.1594181  0.4198117  0.0000
#> treat-0      0.4287845  0.0275311  0.3748245  0.4827444  0.0000
#> differenceG -0.1391696  0.0723524 -0.2809777  0.0026385  0.0544
#> 
#> Average Treatment effects (double robust) :
#>               Estimate   Std.Err      2.5%     97.5% P-value
#> treat-1       0.254948  0.065820  0.125942  0.383953  0.0001
#> treat-0       0.428414  0.027638  0.374244  0.482583  0.0000
#> differenceDR -0.173466  0.071154 -0.312925 -0.034007  0.0148
#> 
#> Average Treatment effects on Treated/Non-Treated (DR) :
#>      Estimate   Std.Err      2.5%     97.5% P-value
#> ATT -0.141215  0.079374 -0.296786  0.014355  0.0752
#> ATC -0.389833  0.052868 -0.493452 -0.286214  0.0000
#> 
#>