Average Treatment effect for censored competing risks data using Binomial Regression

Under the standard causal assumptions we can estimate the average treatment effect E(Y(1) - Y(0)). We need Consistency, ignorability ( Y(1), Y(0) indep A given X), and positivity.

binregATE(
  formula,
  data,
  cause = 1,
  time = NULL,
  beta = NULL,
  treat.model = ~+1,
  cens.model = ~+1,
  offset = NULL,
  weights = NULL,
  cens.weights = NULL,
  se = TRUE,
  kaplan.meier = TRUE,
  cens.code = 0,
  no.opt = FALSE,
  method = "nr",
  augmentation = NULL,
  outcome = c("cif", "rmst", "rmst-cause"),
  model = "exp",
  Ydirect = NULL,
  ...
)

Arguments

formula

formula with outcome (see coxph)

data

data frame

cause

cause of interest

time

time of interest

beta

starting values

treat.model

logistic treatment model given covariates

cens.model

only stratified cox model without covariates

offset

offsets for partial likelihood

weights

for score equations

cens.weights

censoring weights

se

to compute se's with IPCW adjustment, otherwise assumes that IPCW weights are known

kaplan.meier

uses Kaplan-Meier for IPCW in contrast to exp(-Baseline)

cens.code

gives censoring code

no.opt

to not optimize

method

for optimization

augmentation

to augment binomial regression

outcome

can do CIF regression "cif"=F(t|X), "rmst"=E( min(T, t) | X) , or "rmst-cause"=E( I(epsilon==cause) ( t - mint(T,t)) ) | X)

model

possible exp model for E( min(T, t) | X)=exp(X^t beta) , or E( I(epsilon==cause) ( t - mint(T,t)) ) | X)=exp(X^t beta)

Ydirect

use this Y instead of outcome constructed inside the program (e.g. I(T< t, epsilon=1)), then uses IPCW vesion of the Y, set outcome to "rmst" to fit using the model specified by model

...

Additional arguments to lower level funtions

Details

The first covariate in the specification of the competing risks regression model must be the treatment effect that is a factor. If the factor has more than two levels then it uses the mlogit for propensity score modelling. If there are no censorings this is the same as ordinary logistic regression modelling.

Estimates the ATE using the the standard binary double robust estimating equations that are IPCW censoring adjusted. Rather than binomial regression we also consider a IPCW weighted version of standard logistic regression logitIPCWATE.

The original version of the program with only binary treatment binregATEbin take binary-numeric as input for the treatment, and also computes the ATT and ATC, average treatment effect on the treated (ATT), E(Y(1) - Y(0) | A=1), and non-treated, respectively. Experimental version.

Author

Thomas Scheike

Examples

data(bmt)
dfactor(bmt)  <-  ~.

brs <- binregATE(Event(time,cause)~tcell.f+platelet+age,bmt,time=50,cause=1,
  treat.model=tcell.f~platelet+age)
summary(brs)
#> 
#>    n events
#>  408    160
#> 
#>  408 clusters
#> coeffients:
#>              Estimate   Std.Err      2.5%     97.5% P-value
#> (Intercept) -0.195296  0.131037 -0.452123  0.061532  0.1361
#> tcell.f1    -0.637346  0.359010 -1.340993  0.066301  0.0759
#> platelet    -0.351900  0.247162 -0.836329  0.132530  0.1545
#> age          0.419785  0.106904  0.210258  0.629312  0.0001
#> 
#> exp(coeffients):
#>             Estimate    2.5%  97.5%
#> (Intercept)  0.82259 0.63628 1.0635
#> tcell.f1     0.52869 0.26159 1.0685
#> platelet     0.70335 0.43330 1.1417
#> age          1.52163 1.23400 1.8763
#> 
#> Average Treatment effects (G-formula) :
#>             Estimate    Std.Err       2.5%      97.5% P-value
#> treat0     0.4287845  0.0275311  0.3748245  0.4827444  0.0000
#> treat1     0.2896149  0.0664282  0.1594181  0.4198117  0.0000
#> treat:1-0 -0.1391696  0.0723524 -0.2809777  0.0026385  0.0544
#> 
#> Average Treatment effects (double robust) :
#>            Estimate   Std.Err      2.5%     97.5% P-value
#> treat0     0.428414  0.027645  0.374230  0.482597  0.0000
#> treat1     0.254948  0.064321  0.128882  0.381014  0.0001
#> treat:1-0 -0.173466  0.069774 -0.310220 -0.036712  0.0129
#> 
#> 

brsi <- binregATE(Event(time,cause)~tcell.f+tcell.f*platelet+tcell.f*age,bmt,time=50,cause=1,
  treat.model=tcell.f~platelet+age)
summary(brsi)
#> 
#>    n events
#>  408    160
#> 
#>  408 clusters
#> coeffients:
#>                     Estimate    Std.Err       2.5%      97.5% P-value
#> (Intercept)       -0.1639029  0.1330757 -0.4247265  0.0969207  0.2181
#> tcell.f1          -0.9605103  0.4940801 -1.9288895  0.0078688  0.0519
#> platelet          -0.4729820  0.2709874 -1.0041074  0.0581435  0.0809
#> age                0.4267673  0.1116644  0.2079091  0.6456256  0.0001
#> tcell.f1:platelet  0.7995261  0.6966528 -0.5658883  2.1649406  0.2511
#> tcell.f1:age      -0.1298374  0.4266685 -0.9660922  0.7064175  0.7609
#> 
#> exp(coeffients):
#>                   Estimate    2.5%  97.5%
#> (Intercept)        0.84882 0.65395 1.1018
#> tcell.f1           0.38270 0.14531 1.0079
#> platelet           0.62314 0.36637 1.0599
#> age                1.53230 1.23110 1.9072
#> tcell.f1:platelet  2.22449 0.56786 8.7141
#> tcell.f1:age       0.87824 0.38057 2.0267
#> 
#> Average Treatment effects (G-formula) :
#>            Estimate   Std.Err      2.5%     97.5% P-value
#> treat0     0.427926  0.027581  0.373868  0.481985  0.0000
#> treat1     0.269269  0.069599  0.132858  0.405680  0.0001
#> treat:1-0 -0.158657  0.074720 -0.305106 -0.012208  0.0337
#> 
#> Average Treatment effects (double robust) :
#>            Estimate   Std.Err      2.5%     97.5% P-value
#> treat0     0.428394  0.027646  0.374208  0.482579  0.0000
#> treat1     0.257193  0.066194  0.127455  0.386931  0.0001
#> treat:1-0 -0.171201  0.071541 -0.311419 -0.030982  0.0167
#> 
#>