Split a data set and run function from timereg and aggregate

Split a data set and run function of cox-aalen type and aggregate results

divide.conquer.timereg(func = NULL, data, size, ...)

Arguments

func
called function

data
data-frame

size
size of splits

...
Additional arguments to lower level functions

Examples

library(timereg)
data(TRACE)
a <- divide.conquer.timereg(prop.odds,TRACE,
                            formula=Event(time,status==9)~chf+vf+age,n.sim=0,size=200)
coef(a)

#>      Coef.      SE Robust SE D2log(L)^-1     z    P-val lower2.5% upper97.5%
#> chf 1.0700 0.09920    0.0983           1 10.90 0.00e+00    0.8760     1.2600
#> vf  1.5400 0.20100    0.2280           1  6.77 1.28e-11    1.1500     1.9300
#> age 0.0735 0.00529    0.0052           1 14.10 0.00e+00    0.0631     0.0839

a2 <- divide.conquer.timereg(prop.odds,TRACE,
                             formula=Event(time,status==9)~chf+vf+age,n.sim=0,size=500)
coef(a2)

#>      Coef.     SE Robust SE D2log(L)^-1     z    P-val lower2.5% upper97.5%
#> chf 1.0900 0.0982   0.09820           1 11.10 0.00e+00    0.8980     1.2800
#> vf  1.4000 0.1940   0.21500           1  6.53 6.52e-11    1.0200     1.7800
#> age 0.0738 0.0053   0.00522           1 14.10 0.00e+00    0.0634     0.0842

if (interactive()) {
par(mfrow=c(1,1))
plot(a,xlim=c(0,8),ylim=c(0,0.01))
par(new=TRUE)
plot(a2,xlim=c(0,8),ylim=c(0,0.01))
}