list, head, print, tail

listing for data frames

dprint(data, y = NULL, n = 0, ..., x = NULL)

Arguments

data

if x is formula or names for data frame then data frame is needed.

y

name of variable, or fomula, or names of variables on data frame.

n

Index of observations to print (default c(1:nfirst, n-nlast:nlast)

...

Optional additional arguments (nfirst,nlast, and print options)

x

possible group variable

Author

Klaus K. Holst and Thomas Scheike

Examples

n <- 20
m <- lava::lvm(letters)
d <- lava::sim(m,n)

dlist(d,~a+b+c)
#>     a          b        c       
#> 1   -1.949e+00 -0.4087  -0.8903 
#> 2   -5.714e-02  0.3128   0.2369 
#> 3   -7.441e-01 -0.8726   0.9492 
#> 4    9.105e-05  0.4847  -1.1293 
#> 5    5.196e-01  0.3983  -1.4491 
#> ---                             
#> 16   1.7751    -3.43372  0.77383
#> 17  -1.3711     0.03872  0.25082
#> 18  -0.4775    -1.04017 -0.20258
#> 19  -1.7045     0.32440  0.21948
#> 20   0.4772    -1.08441  0.05527
#> attr(,"class")
#> [1] matrix array 
dlist(d,~a+b+c|a<0 & b>0)
#>    a        b       c      
#> 2  -0.05714 0.31284  0.2369
#> 9  -0.14127 0.34328 -0.2589
#> 13 -1.53031 0.32921 -1.1322
#> 17 -1.37114 0.03872  0.2508
#> 19 -1.70449 0.32440  0.2195
#> attr(,"class")
#> [1] matrix array 
## listing all : 
dlist(d,~a+b+c|a<0 & b>0,n=0)
#>              a          b          c
#> 2  -0.05714444 0.31284359  0.2368752
#> 9  -0.14126613 0.34328457 -0.2589379
#> 13 -1.53030617 0.32920793 -1.1321848
#> 17 -1.37114395 0.03871623  0.2508211
#> 19 -1.70449345 0.32440488  0.2194847
dlist(d,a+b+c~I(d>0)|a<0 & b>0)
#> I(d > 0): TRUE
#>    a        b       c      
#> 2  -0.05714 0.31284  0.2369
#> 9  -0.14127 0.34328 -0.2589
#> 13 -1.53031 0.32921 -1.1322
#> 17 -1.37114 0.03872  0.2508
#> 19 -1.70449 0.32440  0.2195
#> attr(,"class")
#> [1] matrix array 
dlist(d,.~I(d>0)|a<0 & b>0)
#> I(d > 0): TRUE
#>    a        b       c       d       e        f       g        h        i      
#> 2  -0.05714 0.31284  0.2369 0.03388 -0.02416  0.1380  0.59850 -0.61864 -0.5969
#> 9  -0.14127 0.34328 -0.2589 0.05180 -1.04807 -1.5899 -0.01145 -0.61701 -0.9026
#> 13 -1.53031 0.32921 -1.1322 0.59971  1.29546  0.7085  1.20921 -0.03558  0.7004
#> 17 -1.37114 0.03872  0.2508 0.08427 -0.77718  1.0119 -0.05549  0.61542 -1.9414
#> 19 -1.70449 0.32440  0.2195 0.33252  0.60908 -0.6864  1.22874  1.08447 -1.3752
#>    j       k        l         m         n        o       p        q      
#> 2  -0.2507  1.63568 -0.623141 -1.066885 -0.20997 -1.3279 -1.29936 -0.9356
#> 9   0.8315  1.52224  0.936547 -0.008779 -1.31368  0.1267 -0.30737  0.4941
#> 13  1.3712 -0.76132 -0.180068  0.265188  0.13694 -0.1373 -0.08348 -0.4199
#> 17  0.1863  0.19538  0.001643 -0.879673 -0.04513 -0.2564  0.63322 -0.7964
#> 19 -0.9703 -0.08724  0.045836  1.413402 -0.73060  0.1288  1.87076 -0.4286
#>    r       s       t        u        v       w        x       y       z      
#> 2   1.6179  0.4369 -0.22109  1.75752 -1.8643  1.31583  1.0670 -0.5685  1.0239
#> 9  -0.7344  2.0857 -1.26964 -0.02934  1.1184 -0.42530  0.5887 -0.3829 -0.4382
#> 13  1.0826  1.4996  0.53322  0.09417  0.3304 -1.33542 -0.3957 -0.4936 -0.4721
#> 17  0.6981  0.5224 -0.07473  0.38958 -3.3255  0.48166 -0.2818  1.3355 -1.4386
#> 19 -0.1806 -1.0727 -0.37560 -1.09867  1.5641 -0.04462  0.4747 -0.8430 -1.9859
#> attr(,"class")
#> [1] matrix array 
dlist(d,~a+b+c|a<0 & b>0, nlast=0)
#>    a        b       c      
#> 2  -0.05714 0.31284  0.2369
#> 9  -0.14127 0.34328 -0.2589
#> 13 -1.53031 0.32921 -1.1322
#> 17 -1.37114 0.03872  0.2508
#> 19 -1.70449 0.32440  0.2195
#> attr(,"class")
#> [1] matrix array 
dlist(d,~a+b+c|a<0 & b>0, nfirst=3, nlast=3)
#>    a        b       c      
#> 2  -0.05714 0.31284  0.2369
#> 9  -0.14127 0.34328 -0.2589
#> 13 -1.53031 0.32921 -1.1322
#> 17 -1.37114 0.03872  0.2508
#> 19 -1.70449 0.32440  0.2195
#> attr(,"class")
#> [1] matrix array 
dlist(d,~a+b+c|a<0 & b>0, 1:5)
#>    a        b       c      
#> 2  -0.05714 0.31284  0.2369
#> 9  -0.14127 0.34328 -0.2589
#> 13 -1.53031 0.32921 -1.1322
#> 17 -1.37114 0.03872  0.2508
#> 19 -1.70449 0.32440  0.2195
#> attr(,"class")
#> [1] matrix array 
dlist(d,~a+b+c|a<0 & b>0, -(5:1))
#>    a        b       c      
#> 2  -0.05714 0.31284  0.2369
#> 9  -0.14127 0.34328 -0.2589
#> 13 -1.53031 0.32921 -1.1322
#> 17 -1.37114 0.03872  0.2508
#> 19 -1.70449 0.32440  0.2195
#> attr(,"class")
#> [1] matrix array 
dlist(d,~a+b+c|a<0 & b>0, list(1:5,50:55,-(5:1)))
#>     a        b       c      
#> 2   -0.05714 0.31284  0.2369
#> 9   -0.14127 0.34328 -0.2589
#> 13  -1.53031 0.32921 -1.1322
#> 17  -1.37114 0.03872  0.2508
#> 19  -1.70449 0.32440  0.2195
#> ---                         
#> ---                         
#> 2   -0.05714 0.31284  0.2369
#> 9   -0.14127 0.34328 -0.2589
#> 13  -1.53031 0.32921 -1.1322
#> 17  -1.37114 0.03872  0.2508
#> 19  -1.70449 0.32440  0.2195
#> attr(,"class")
#> [1] matrix array 
dprint(d,a+b+c ~ I(d>0) |a<0 & b>0, list(1:5,50:55,-(5:1)))
#> I(d > 0): TRUE
#>     a        b       c      
#> 2   -0.05714 0.31284  0.2369
#> 9   -0.14127 0.34328 -0.2589
#> 13  -1.53031 0.32921 -1.1322
#> 17  -1.37114 0.03872  0.2508
#> 19  -1.70449 0.32440  0.2195
#> ---                         
#> ---                         
#> 2   -0.05714 0.31284  0.2369
#> 9   -0.14127 0.34328 -0.2589
#> 13  -1.53031 0.32921 -1.1322
#> 17  -1.37114 0.03872  0.2508
#> 19  -1.70449 0.32440  0.2195
#> attr(,"class")
#> [1] matrix array