list, head, print, tail

listing for data frames

dprint(data, y = NULL, n = 0, ..., x = NULL)

Arguments

data

if x is formula or names for data frame then data frame is needed.

y

name of variable, or fomula, or names of variables on data frame.

n

Index of observations to print (default c(1:nfirst, n-nlast:nlast)

...

Optional additional arguments (nfirst,nlast, and print options)

x

possible group variable

Author

Klaus K. Holst and Thomas Scheike

Examples

m <- lava::lvm(letters)
d <- lava::sim(m, 20)

dlist(d,~a+b+c)
#>     a        b       c        
#> 1   -0.57713  1.3864  0.2035  
#> 2   -0.80592 -1.3047  0.8770  
#> 3    1.22148  0.8083 -0.8739  
#> 4    0.03472  0.1451 -0.2107  
#> 5   -2.55785  0.8355  0.5970  
#> ---                           
#> 16   1.0592   0.1010 -0.004013
#> 17   1.0293  -0.1649  0.582503
#> 18  -0.3133  -0.7004  0.170292
#> 19  -0.9958   1.8680  0.055593
#> 20  -0.6279  -0.5425  1.141641
dlist(d,~a+b+c|a<0 & b>0)
#>    a       b       c       
#> 1  -0.5771 1.38640  0.20349
#> 5  -2.5578 0.83551  0.59703
#> 7  -1.4403 1.84722 -1.36392
#> 12 -0.6864 0.34647  1.12117
#> 13 -0.4690 0.03698  0.21409
#> 14 -0.8690 0.24112  0.85016
#> 19 -0.9958 1.86798  0.05559
## listing all : 
dlist(d,~a+b+c|a<0 & b>0,n=0)
#>             a          b           c
#> 1  -0.5771254 1.38639858  0.20349331
#> 5  -2.5578479 0.83551163  0.59702674
#> 7  -1.4402667 1.84721571 -1.36391844
#> 12 -0.6863567 0.34647417  1.12117388
#> 13 -0.4689727 0.03698312  0.21408719
#> 14 -0.8689645 0.24111853  0.85016454
#> 19 -0.9958388 1.86798026  0.05559288
dlist(d,a+b+c~I(d>0)|a<0 & b>0)
#> I(d > 0): FALSE
#>   a     b     c     
#> 7 -1.44 1.847 -1.364
#> ------------------------------------------------------------ 
#> I(d > 0): TRUE
#>    a       b       c      
#> 1  -0.5771 1.38640 0.20349
#> 5  -2.5578 0.83551 0.59703
#> 12 -0.6864 0.34647 1.12117
#> 13 -0.4690 0.03698 0.21409
#> 14 -0.8690 0.24112 0.85016
#> 19 -0.9958 1.86798 0.05559
dlist(d,.~I(d>0)|a<0 & b>0)
#> I(d > 0): FALSE
#>   a     b     c      d       e      f      g      h       i      j      k    
#> 7 -1.44 1.847 -1.364 -0.5632 -1.316 0.3615 0.5216 -0.1331 0.4985 -1.102 1.115
#>   l      m      n      o      p      q     r      s       t       u      v     
#> 7 0.7383 -1.494 0.1968 0.3385 0.4008 0.491 0.6065 -0.3661 -0.4159 -1.657 -2.589
#>   w      x       y     z    
#> 7 -1.496 -0.5205 -1.21 1.693
#> ------------------------------------------------------------ 
#> I(d > 0): TRUE
#>    a       b       c       d      e        f       g         h       i       
#> 1  -0.5771 1.38640 0.20349 1.4098  0.04847 -1.9705  0.658111  0.4043 -0.52642
#> 5  -2.5578 0.83551 0.59703 0.3855  0.51960 -0.1985  0.821824 -0.8275  1.08230
#> 12 -0.6864 0.34647 1.12117 0.5538  1.96098  1.5414  0.194344  0.1269 -0.04678
#> 13 -0.4690 0.03698 0.21409 0.1177 -0.46321  0.9789  0.936691 -0.3200 -0.94240
#> 14 -0.8690 0.24112 0.85016 0.9213 -0.86465 -0.3636 -0.253180  0.5867 -0.16376
#> 19 -0.9958 1.86798 0.05559 0.9732  0.44719 -1.1159  0.001865  0.2458  0.09711
#>    j       k        l        m       n         o        p       q      
#> 1   0.3942  1.38818  0.17237 -0.5659 -0.007628  0.04385 -0.5305 -0.4197
#> 5   0.1518 -0.07623 -0.36970  1.1151 -0.299895 -0.57605  0.5141 -0.3771
#> 12 -1.2743 -0.05771 -0.68983 -1.2597  0.613626 -0.33646  0.8328 -0.9099
#> 13  0.7546  0.14783  0.04554 -0.4604  0.017346  0.87833  1.4594  1.3654
#> 14 -0.4069 -0.14879  0.82880 -1.3357 -0.700748  1.24798  1.5513 -0.5203
#> 19  0.5729 -0.45289 -0.34918 -0.1733 -0.341316 -0.20052  0.6914 -1.6135
#>    r         s       t       u         v       w        x       y       
#> 1   1.231414  1.1565 -0.6686 -0.699676  0.6825  0.53973 2.29360  0.97959
#> 5  -1.090981 -0.7229  1.5300 -1.217620 -0.7953 -0.30082 0.78596  0.60130
#> 12 -0.237471 -1.4424 -0.3349  0.918788  0.2956 -0.19642 0.33206 -0.29644
#> 13 -0.009664 -0.5202  0.9787  0.338243  0.1452  0.45054 0.47294  0.03932
#> 14  1.532874 -0.4943  0.6186 -0.008522 -0.2978  1.73364 0.39648  2.58308
#> 19  2.351449  0.6613  0.7215  1.700028 -0.4123  0.02187 0.06938 -0.73137
#>    z       
#> 1   0.65615
#> 5  -0.02632
#> 12 -0.13111
#> 13 -0.99994
#> 14 -0.06770
#> 19  1.57627
dlist(d,~a+b+c|a<0 & b>0, nlast=0)
#>    a       b       c      
#> 1  -0.5771 1.38640  0.2035
#> 5  -2.5578 0.83551  0.5970
#> 7  -1.4403 1.84722 -1.3639
#> 12 -0.6864 0.34647  1.1212
#> 13 -0.4690 0.03698  0.2141
dlist(d,~a+b+c|a<0 & b>0, nfirst=3, nlast=3)
#>     a       b       c      
#> 1   -0.5771 1.3864   0.2035
#> 5   -2.5578 0.8355   0.5970
#> 7   -1.4403 1.8472  -1.3639
#> ---                        
#> 13  -0.4690 0.03698 0.21409
#> 14  -0.8690 0.24112 0.85016
#> 19  -0.9958 1.86798 0.05559
dlist(d,~a+b+c|a<0 & b>0, 1:5)
#>    a       b       c      
#> 1  -0.5771 1.38640  0.2035
#> 5  -2.5578 0.83551  0.5970
#> 7  -1.4403 1.84722 -1.3639
#> 12 -0.6864 0.34647  1.1212
#> 13 -0.4690 0.03698  0.2141
dlist(d,~a+b+c|a<0 & b>0, -(5:1))
#>    a       b       c       
#> 7  -1.4403 1.84722 -1.36392
#> 12 -0.6864 0.34647  1.12117
#> 13 -0.4690 0.03698  0.21409
#> 14 -0.8690 0.24112  0.85016
#> 19 -0.9958 1.86798  0.05559
dlist(d,~a+b+c|a<0 & b>0, list(1:5,50:55,-(5:1)))
#>     a       b       c       
#> 1   -0.5771 1.38640  0.2035 
#> 5   -2.5578 0.83551  0.5970 
#> 7   -1.4403 1.84722 -1.3639 
#> 12  -0.6864 0.34647  1.1212 
#> 13  -0.4690 0.03698  0.2141 
#> ---                         
#> ---                         
#> 7   -1.4403 1.84722 -1.36392
#> 12  -0.6864 0.34647  1.12117
#> 13  -0.4690 0.03698  0.21409
#> 14  -0.8690 0.24112  0.85016
#> 19  -0.9958 1.86798  0.05559
dprint(d,a+b+c ~ I(d>0) |a<0 & b>0, list(1:5,50:55,-(5:1)))
#> I(d > 0): FALSE
#>     a     b     c     
#> 7   -1.44 1.847 -1.364
#> ---                   
#> ---                   
#> 7   -1.44 1.847 -1.364
#> ------------------------------------------------------------ 
#> I(d > 0): TRUE
#>     a       b       c      
#> 1   -0.5771 1.38640 0.2035 
#> 5   -2.5578 0.83551 0.5970 
#> 12  -0.6864 0.34647 1.1212 
#> 13  -0.4690 0.03698 0.2141 
#> 14  -0.8690 0.24112 0.8502 
#> ---                        
#> ---                        
#> 5   -2.5578 0.83551 0.59703
#> 12  -0.6864 0.34647 1.12117
#> 13  -0.4690 0.03698 0.21409
#> 14  -0.8690 0.24112 0.85016
#> 19  -0.9958 1.86798 0.05559