list, head, print, tail

listing for data frames

dprint(data, y = NULL, n = 0, ..., x = NULL)

Arguments

data
if x is formula or names for data frame then data frame is needed.

y
name of variable, or fomula, or names of variables on data frame.

n
Index of observations to print (default c(1:nfirst, n-nlast:nlast)

...
Optional additional arguments (nfirst,nlast, and print options)

x
possible group variable

Author

Klaus K. Holst and Thomas Scheike

Examples

n <- 20
m <- lava::lvm(letters)
d <- lava::sim(m,n)

dlist(d,~a+b+c)

#>     a        b       c       
#> 1   -1.84383 -0.4530 -0.7494 
#> 2    0.05678  0.3284 -0.2295 
#> 3    0.20924 -0.1455  1.1472 
#> 4    1.62813 -0.6004  3.0233 
#> 5    1.86731 -0.2218  0.6702 
#> ---                          
#> 16  -0.05112 -0.536  -0.05565
#> 17   0.11700  1.137  -0.42263
#> 18  -0.84498  0.986  -2.10878
#> 19   1.72061  1.598   1.03400
#> 20  -0.98638 -1.690   0.42084
#> attr(,"class")
#> [1] matrix array 

dlist(d,~a+b+c|a<0 & b>0)

#>    a       b      c      
#> 8  -2.5006 0.2945 -0.6721
#> 15 -0.2636 0.6685 -0.4457
#> 18 -0.8450 0.9860 -2.1088
#> attr(,"class")
#> [1] matrix array 

## listing all : 
dlist(d,~a+b+c|a<0 & b>0,n=0)

#>             a         b          c
#> 8  -2.5005946 0.2945457 -0.6720881
#> 15 -0.2635549 0.6685403 -0.4457079
#> 18 -0.8449837 0.9860376 -2.1087776

dlist(d,a+b+c~I(d>0)|a<0 & b>0)

#> I(d > 0): FALSE
#>    a       b      c      
#> 8  -2.5006 0.2945 -0.6721
#> 15 -0.2636 0.6685 -0.4457
#> attr(,"class")
#> [1] matrix array 
#> ------------------------------------------------------------ 
#> I(d > 0): TRUE
#>    a      b     c     
#> 18 -0.845 0.986 -2.109
#> attr(,"class")
#> [1] matrix array 

dlist(d,.~I(d>0)|a<0 & b>0)

#> I(d > 0): FALSE
#>    a       b      c       d       e       f        g      h       i     
#> 8  -2.5006 0.2945 -0.6721 -0.3044  0.2865 -1.21571 1.5568 0.04573 0.5221
#> 15 -0.2636 0.6685 -0.4457 -0.1470 -1.0489  0.09333 0.1512 1.41278 0.3651
#>    j       k        l       m       n     o      p       q        r       
#> 8  -0.1554  0.27277 -0.2019 -2.4884 0.758 -2.639 -0.1937 -0.55268  0.00341
#> 15  0.7464 -0.07459  0.1186 -0.7062 1.379 -1.167  1.0038 -0.07095 -1.08212
#>    s       t       u       v       w       x       y      z      
#> 8  -0.1345  1.7241  0.2622 -2.2236 -0.4376  0.1059  0.490  0.3116
#> 15  0.9054 -0.4916 -0.7992 -0.1184  0.4878 -0.7961 -1.786 -0.4836
#> attr(,"class")
#> [1] matrix array 
#> ------------------------------------------------------------ 
#> I(d > 0): TRUE
#>    a      b     c      d     e      f       g       h       i       j     
#> 18 -0.845 0.986 -2.109 1.062 0.2796 -0.1988 0.02636 -0.8364 -0.1294 0.1421
#>    k     l     m     n      o     p      q      r       s       t       u     
#> 18 0.141 1.366 -1.06 0.8933 -1.16 0.1766 0.5121 -0.4789 -0.1393 -0.2209 0.4415
#>    v      w      x       y       z      
#> 18 0.2771 0.4051 -0.9893 -0.8414 -0.5754
#> attr(,"class")
#> [1] matrix array 

dlist(d,~a+b+c|a<0 & b>0, nlast=0)

#>    a       b      c      
#> 8  -2.5006 0.2945 -0.6721
#> 15 -0.2636 0.6685 -0.4457
#> 18 -0.8450 0.9860 -2.1088
#> attr(,"class")
#> [1] matrix array 

dlist(d,~a+b+c|a<0 & b>0, nfirst=3, nlast=3)

#>    a       b      c      
#> 8  -2.5006 0.2945 -0.6721
#> 15 -0.2636 0.6685 -0.4457
#> 18 -0.8450 0.9860 -2.1088
#> attr(,"class")
#> [1] matrix array 

dlist(d,~a+b+c|a<0 & b>0, 1:5)

#>    a       b      c      
#> 8  -2.5006 0.2945 -0.6721
#> 15 -0.2636 0.6685 -0.4457
#> 18 -0.8450 0.9860 -2.1088
#> attr(,"class")
#> [1] matrix array 

dlist(d,~a+b+c|a<0 & b>0, -(5:1))

#>    a       b      c      
#> 8  -2.5006 0.2945 -0.6721
#> 15 -0.2636 0.6685 -0.4457
#> 18 -0.8450 0.9860 -2.1088
#> attr(,"class")
#> [1] matrix array 

dlist(d,~a+b+c|a<0 & b>0, list(1:5,50:55,-(5:1)))

#>     a       b      c      
#> 8   -2.5006 0.2945 -0.6721
#> 15  -0.2636 0.6685 -0.4457
#> 18  -0.8450 0.9860 -2.1088
#> ---                       
#> ---                       
#> 8   -2.5006 0.2945 -0.6721
#> 15  -0.2636 0.6685 -0.4457
#> 18  -0.8450 0.9860 -2.1088
#> attr(,"class")
#> [1] matrix array 

dprint(d,a+b+c ~ I(d>0) |a<0 & b>0, list(1:5,50:55,-(5:1)))

#> I(d > 0): FALSE
#>     a       b      c      
#> 8   -2.5006 0.2945 -0.6721
#> 15  -0.2636 0.6685 -0.4457
#> ---                       
#> ---                       
#> 8   -2.5006 0.2945 -0.6721
#> 15  -0.2636 0.6685 -0.4457
#> attr(,"class")
#> [1] matrix array 
#> ------------------------------------------------------------ 
#> I(d > 0): TRUE
#>     a      b     c     
#> 18  -0.845 0.986 -2.109
#> ---                    
#> ---                    
#> 18  -0.845 0.986 -2.109
#> attr(,"class")
#> [1] matrix array