Fits two-stage binomial for describing depdendence in binomial data using marginals that are on logistic form using the binomial.twostage funcion, but call is different and easier and the data manipulation is build into the function. Useful in particular for family design data.

If clusters contain more than two times, the algoritm uses a compososite likelihood based on the pairwise bivariate models.

easy.binomial.twostage(
  margbin = NULL,
  data = parent.frame(),
  method = "nr",
  response = "response",
  id = "id",
  Nit = 60,
  detail = 0,
  silent = 1,
  weights = NULL,
  control = list(),
  theta = NULL,
  theta.formula = NULL,
  desnames = NULL,
  deshelp = 0,
  var.link = 1,
  iid = 1,
  step = 1,
  model = "plackett",
  marginal.p = NULL,
  strata = NULL,
  max.clust = NULL,
  se.clusters = NULL
)

Arguments

margbin
Marginal binomial model

data
data frame

method
Scoring method

response
name of response variable in data frame

id
name of cluster variable in data frame

Nit
Number of iterations

detail
Detail for more output for iterations

silent
Debug information

weights
Weights for log-likelihood, can be used for each type of outcome in 2x2 tables.

control
Optimization arguments

theta
Starting values for variance components

theta.formula
design for depedence, either formula or design function

desnames
names for dependence parameters

deshelp
if 1 then prints out some data sets that are used, on on which the design function operates

var.link
Link function for variance

iid
Calculate i.i.d. decomposition

step
Step size

model
model

marginal.p
vector of marginal probabilities

strata
strata for fitting

max.clust
max clusters used for i.i.d. decompostion

se.clusters
clusters for iid decomposition for roubst standard errors

Details

The reported standard errors are based on the estimated information from the likelihood assuming that the marginals are known. This gives correct standard errors in the case of the plackett distribution (OR model for dependence), but incorrect for the clayton-oakes types model. The OR model is often known as the ALR model. Our fitting procedures gives correct standard errors due to the ortogonality and is fast.

Examples

data(twinstut)
twinstut0 <- subset(twinstut, tvparnr<2300000)
twinstut <- twinstut0
twinstut$binstut <- (twinstut$stutter=="yes")*1
theta.des <- model.matrix( ~-1+factor(zyg),data=twinstut)
margbin <- glm(binstut~factor(sex)+age,data=twinstut,family=binomial())
bin <- binomial.twostage(margbin,data=twinstut,var.link=1,
                         clusters=twinstut$tvparnr,theta.des=theta.des,detail=0,
                         method="nr")
summary(bin)

#> Dependence parameter for Odds-Ratio (Plackett) model 
#> With log-link 

#> $estimates
#>                   theta        se
#> factor(zyg)dz 0.5278527 0.2396796
#> factor(zyg)mz 3.4850037 0.1864190
#> factor(zyg)os 0.7802940 0.2894394
#> 
#> $or
#>               Estimate Std.Err    2.5%  97.5%   P-value
#> factor(zyg)dz    1.695  0.4063  0.8989  2.492 3.016e-05
#> factor(zyg)mz   32.623  6.0815 20.7031 44.542 8.128e-08
#> factor(zyg)os    2.182  0.6316  0.9442  3.420 5.504e-04
#> 
#> $type
#> [1] "plackett"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

lava::estimate(coef=bin$theta,vcov=bin$var.theta,f=function(p) exp(p))

#>      Estimate Std.Err    2.5%  97.5%   P-value
#> NA      1.695  0.4063  0.8989  2.492 3.016e-05
#> NA.1   32.623  6.0815 20.7031 44.542 8.128e-08
#> NA.2    2.182  0.6316  0.9442  3.420 5.504e-04

twinstut$cage <- scale(twinstut$age)
theta.des <- model.matrix( ~-1+factor(zyg)+cage,data=twinstut)
bina <- binomial.twostage(margbin,data=twinstut,var.link=1,
                         clusters=twinstut$tvparnr,theta.des=theta.des,detail=0)
summary(bina)

#> Dependence parameter for Odds-Ratio (Plackett) model 
#> With log-link 

#> $estimates
#>                     theta        se
#> factor(zyg)dz  0.52618528 0.2404053
#> factor(zyg)mz  3.47580396 0.1929538
#> factor(zyg)os  0.78180443 0.2896335
#> cage          -0.02603244 0.1366320
#> 
#> $or
#>               Estimate Std.Err    2.5%  97.5%   P-value
#> factor(zyg)dz   1.6925  0.4069  0.8950  2.490 3.187e-05
#> factor(zyg)mz  32.3238  6.2370 20.0995 44.548 2.188e-07
#> factor(zyg)os   2.1854  0.6330  0.9448  3.426 5.551e-04
#> cage            0.9743  0.1331  0.7134  1.235 2.500e-13
#> 
#> $type
#> [1] "plackett"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

theta.des <- model.matrix( ~-1+factor(zyg)+factor(zyg)*cage,data=twinstut)
bina <- binomial.twostage(margbin,data=twinstut,var.link=1,
                         clusters=twinstut$tvparnr,theta.des=theta.des)
summary(bina)

#> Dependence parameter for Odds-Ratio (Plackett) model 
#> With log-link 

#> $estimates
#>                         theta        se
#> factor(zyg)dz       0.5184191 0.2425122
#> factor(zyg)mz       3.5157878 0.1983717
#> factor(zyg)os       0.7806223 0.2901748
#> cage               -0.1082647 0.2640090
#> factor(zyg)mz:cage  0.1939135 0.3225548
#> factor(zyg)os:cage -0.1284699 0.4095734
#> 
#> $or
#>                    Estimate Std.Err    2.5%  97.5%   P-value
#> factor(zyg)dz        1.6794  0.4073  0.8811  2.478 3.732e-05
#> factor(zyg)mz       33.6424  6.6737 20.5622 46.723 4.630e-07
#> factor(zyg)os        2.1828  0.6334  0.9414  3.424 5.685e-04
#> cage                 0.8974  0.2369  0.4330  1.362 1.520e-04
#> factor(zyg)mz:cage   1.2140  0.3916  0.4465  1.981 1.934e-03
#> factor(zyg)os:cage   0.8794  0.3602  0.1735  1.585 1.462e-02
#> 
#> $type
#> [1] "plackett"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

out <- easy.binomial.twostage(stutter~factor(sex)+age,data=twinstut,
                              response="binstut",id="tvparnr",var.link=1,
                                  theta.formula=~-1+factor(zyg1))
summary(out)

#> Dependence parameter for Odds-Ratio (Plackett) model 
#> With log-link 

#> $estimates
#>                    theta        se
#> factor(zyg1)dz 0.5278527 0.2396796
#> factor(zyg1)mz 3.4850037 0.1864190
#> factor(zyg1)os 0.7802940 0.2894394
#> 
#> $or
#>                Estimate Std.Err    2.5%  97.5%   P-value
#> factor(zyg1)dz    1.695  0.4063  0.8989  2.492 3.016e-05
#> factor(zyg1)mz   32.623  6.0815 20.7031 44.542 8.128e-08
#> factor(zyg1)os    2.182  0.6316  0.9442  3.420 5.504e-04
#> 
#> $type
#> [1] "plackett"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

## refers to zygosity of first subject in eash pair : zyg1
## could also use zyg2 (since zyg2=zyg1 within twinpair's))
desfs <- function(x,num1="zyg1",namesdes=c("mz","dz","os"))
    c(x[num1]=="mz",x[num1]=="dz",x[num1]=="os")*1

out3 <- easy.binomial.twostage(binstut~factor(sex)+age,
                               data=twinstut, response="binstut",id="tvparnr",
                               var.link=1,theta.formula=desfs,
                               desnames=c("mz","dz","os"))
summary(out3)

#> Dependence parameter for Odds-Ratio (Plackett) model 
#> With log-link 

#> $estimates
#>        theta        se
#> mz 3.4850037 0.1864190
#> dz 0.5278527 0.2396796
#> os 0.7802940 0.2894394
#> 
#> $or
#>    Estimate Std.Err    2.5%  97.5%   P-value
#> mz   32.623  6.0815 20.7031 44.542 8.128e-08
#> dz    1.695  0.4063  0.8989  2.492 3.016e-05
#> os    2.182  0.6316  0.9442  3.420 5.504e-04
#> 
#> $type
#> [1] "plackett"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

 ## Reduce Ex.Timings
n <- 10000
set.seed(100)
dd <- simBinFam(n,beta=0.3)
binfam <- fast.reshape(dd,varying=c("age","x","y"))
## mother, father, children  (ordered)
head(binfam)

#>   id      age x y num
#> 1  1 20.44017 0 1   m
#> 2  1 24.17965 0 1   f
#> 3  1 10.63289 0 1  b1
#> 4  1 12.11630 0 1  b2
#> 5  2 28.03707 1 1   m
#> 6  2 25.00649 1 0   f

########### ########### ########### ########### ########### ###########
####  simple analyses of binomial family data
########### ########### ########### ########### ########### ###########
desfs <- function(x,num1="num1",num2="num2")
{
     pp <- 1*(((x[num1]=="m")*(x[num2]=="f"))|(x[num1]=="f")*(x[num2]=="m"))
     pc <- (x[num1]=="m" | x[num1]=="f")*(x[num2]=="b1" | x[num2]=="b2")*1
     cc <- (x[num1]=="b1")*(x[num2]=="b1" | x[num2]=="b2")*1
     c(pp,pc,cc)
}

ud <- easy.binomial.twostage(y~+1,data=binfam,
     response="y",id="id",
     theta.formula=desfs,desnames=c("pp","pc","cc"))
summary(ud)

#> Dependence parameter for Odds-Ratio (Plackett) model 
#> With log-link 

#> $estimates
#>         theta         se
#> pp 0.02714191 0.07881159
#> pc 0.31959351 0.03689402
#> cc 0.65795314 0.06659724
#> 
#> $or
#>    Estimate Std.Err   2.5% 97.5%    P-value
#> pp    1.028 0.08098 0.8688 1.186  6.849e-37
#> pc    1.377 0.05079 1.2770 1.476 8.676e-162
#> cc    1.931 0.12859 1.6788 2.183  5.800e-51
#> 
#> $type
#> [1] "plackett"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

udx <- easy.binomial.twostage(y~+x,data=binfam,
     response="y",id="id",
     theta.formula=desfs,desnames=c("pp","pc","cc"))
summary(udx)

#> Dependence parameter for Odds-Ratio (Plackett) model 
#> With log-link 

#> $estimates
#>         theta         se
#> pp 0.02777621 0.07893634
#> pc 0.32266211 0.03698398
#> cc 0.66038803 0.06680438
#> 
#> $or
#>    Estimate Std.Err   2.5% 97.5%    P-value
#> pp    1.028 0.08116 0.8691 1.187  8.846e-37
#> pc    1.381 0.05107 1.2807 1.481 5.182e-161
#> cc    1.936 0.12930 1.6821 2.189  1.169e-50
#> 
#> $type
#> [1] "plackett"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

########### ########### ########### ########### ########### ###########
####  now allowing parent child POR to be different for mother and father
########### ########### ########### ########### ########### ###########

desfsi <- function(x,num1="num1",num2="num2")
{
    pp <- (x[num1]=="m")*(x[num2]=="f")*1
    mc <- (x[num1]=="m")*(x[num2]=="b1" | x[num2]=="b2")*1
    fc <- (x[num1]=="f")*(x[num2]=="b1" | x[num2]=="b2")*1
    cc <- (x[num1]=="b1")*(x[num2]=="b1" | x[num2]=="b2")*1
    c(pp,mc,fc,cc)
}

udi <- easy.binomial.twostage(y~+1,data=binfam,
     response="y",id="id",
     theta.formula=desfsi,desnames=c("pp","mother-child","father-child","cc"))
summary(udi)

#> Dependence parameter for Odds-Ratio (Plackett) model 
#> With log-link 

#> $estimates
#>                   theta         se
#> pp           0.02714191 0.07881159
#> mother-child 0.52984686 0.05036850
#> father-child 0.09176086 0.05581411
#> cc           0.65795314 0.06659724
#> 
#> $or
#>              Estimate Std.Err   2.5% 97.5%   P-value
#> pp              1.028 0.08098 0.8688 1.186 6.849e-37
#> mother-child    1.699 0.08556 1.5310 1.866 1.024e-87
#> father-child    1.096 0.06118 0.9762 1.216 8.749e-72
#> cc              1.931 0.12859 1.6788 2.183 5.800e-51
#> 
#> $type
#> [1] "plackett"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

##now looking to see if interactions with age or age influences marginal models
##converting factors to numeric to make all involved covariates numeric
##to use desfai2 rather then desfai that works on binfam

nbinfam <- binfam
nbinfam$num <- as.numeric(binfam$num)
head(nbinfam)

#>   id      age x y num
#> 1  1 20.44017 0 1   1
#> 2  1 24.17965 0 1   2
#> 3  1 10.63289 0 1   3
#> 4  1 12.11630 0 1   4
#> 5  2 28.03707 1 1   1
#> 6  2 25.00649 1 0   2

desfsai <- function(x,num1="num1",num2="num2")
{
    pp <- (x[num1]=="m")*(x[num2]=="f")*1
### av age for pp=1 i.e parent pairs
    agepp <- ((as.numeric(x["age1"])+as.numeric(x["age2"]))/2-30)*pp
    mc <- (x[num1]=="m")*(x[num2]=="b1" | x[num2]=="b2")*1
    fc <- (x[num1]=="f")*(x[num2]=="b1" | x[num2]=="b2")*1
    cc <- (x[num1]=="b1")*(x[num2]=="b1" | x[num2]=="b2")*1
    agecc <- ((as.numeric(x["age1"])+as.numeric(x["age2"]))/2-12)*cc
    c(pp,agepp,mc,fc,cc,agecc)
}

desfsai2 <- function(x,num1="num1",num2="num2")
{
    pp <- (x[num1]==1)*(x[num2]==2)*1
    agepp <- (((x["age1"]+x["age2"]))/2-30)*pp ### av age for pp=1 i.e parent pairs
    mc <- (x[num1]==1)*(x[num2]==3 | x[num2]==4)*1
    fc <- (x[num1]==2)*(x[num2]==3 | x[num2]==4)*1
    cc <- (x[num1]==3)*(x[num2]==3 | x[num2]==4)*1
    agecc <- ((x["age1"]+x["age2"])/2-12)*cc ### av age for children
    c(pp,agepp,mc,fc,cc,agecc)
}

udxai2 <- easy.binomial.twostage(y~+x+age,data=binfam,
     response="y",id="id",
     theta.formula=desfsai,
     desnames=c("pp","pp-age","mother-child","father-child","cc","cc-age"))
summary(udxai2)

#> Dependence parameter for Odds-Ratio (Plackett) model 
#> With log-link 

#> $estimates
#>                    theta         se
#> pp           -0.04132082 0.11936287
#> pp-age       -0.01937892 0.02485374
#> mother-child  0.53543124 0.05066611
#> father-child  0.09288514 0.05583962
#> cc            0.68690859 0.06870483
#> cc-age        0.03718788 0.02336246
#> 
#> $or
#>              Estimate Std.Err   2.5% 97.5%   P-value
#> pp             0.9595 0.11453 0.7350 1.184 5.392e-17
#> pp-age         0.9808 0.02438 0.9330 1.029 0.000e+00
#> mother-child   1.7082 0.08655 1.5386 1.878 1.036e-86
#> father-child   1.0973 0.06127 0.9772 1.217 1.013e-71
#> cc             1.9876 0.13656 1.7199 2.255 5.427e-48
#> cc-age         1.0379 0.02425 0.9904 1.085 0.000e+00
#> 
#> $type
#> [1] "plackett"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"