Wrapper for easy fitting of Clayton-Oakes or bivariate Plackett models for bivariate survival data

Fits two-stage model for describing depdendence in survival data using marginals that are on cox or aalen form using the twostage funcion, but call is different and easier and the data manipulation build into the function. Useful in particular for family design data.

easy.survival.twostage(
  margsurv = NULL,
  data = parent.frame(),
  method = "nlminb",
  status = "status",
  time = "time",
  entry = NULL,
  id = "id",
  Nit = 60,
  detail = 0,
  silent = 1,
  weights = NULL,
  control = list(),
  theta = NULL,
  theta.formula = NULL,
  desnames = NULL,
  deshelp = 0,
  var.link = 1,
  step = 0.5,
  model = "plackett",
  marginal.surv = NULL,
  strata = NULL,
  se.clusters = NULL
)

Arguments

margsurv
model

data
data frame

method
Scoring method

status
Status at exit time

time
Exit time

entry
Entry time

id
name of cluster variable in data frame

Nit
Number of iterations

detail
Detail for more output for iterations

silent
Debug information

weights
Weights for log-likelihood, can be used for each type of outcome in 2x2 tables.

control
Optimization arguments

theta
Starting values for variance components

theta.formula
design for depedence, either formula or design function

desnames
names for dependence parameters

deshelp
if 1 then prints out some data sets that are used, on on which the design function operates

var.link
Link function for variance (exp link)

step
Step size for newton-raphson

model
plackett or clayton-oakes model

marginal.surv
vector of marginal survival probabilities

strata
strata for fitting

se.clusters
clusters for iid decomposition for roubst standard errors

Details

If clusters contain more than two times, the algoritm uses a composite likelihood based on the pairwise bivariate models.

The reported standard errors are based on the estimated information from the likelihood assuming that the marginals are known.

Examples

library(mets)
data("prt",package="mets")
prtsam <- blocksample(prt,idvar="id",1e3,replace=FALSE)
margp <- coxph(Surv(time,status==1)~factor(country),data=prtsam)
fitco <- survival.twostage(margp,data=prtsam,clusters=prtsam$id)
summary(fitco)

#> Dependence parameter for Clayton-Oakes model
#> Variance of Gamma distributed random effects 
#> With log-link 

#> $estimates
#>             log-Coef.        SE         z       P-val Kendall tau         SE
#> dependence1  -1.47177 0.4992257 -2.948105 0.003197286   0.1029455 0.04610236
#> 
#> $vargam
#>             Estimate Std.Err     2.5%  97.5% P-value
#> dependence1   0.2295  0.1146 0.004943 0.4541 0.04517
#> 
#> $type
#> [1] "clayton.oakes"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

des <- model.matrix(~-1+factor(zyg),data=prtsam);
fitco <- survival.twostage(margp,data=prtsam,theta.des=des,clusters=prtsam$id)
summary(fitco)

#> Dependence parameter for Clayton-Oakes model
#> Variance of Gamma distributed random effects 
#> With log-link 

#> $estimates
#>                log-Coef.        SE         z      P-val Kendall tau         SE
#> factor(zyg)DZ -1.8118821 0.7803716 -2.321820 0.02024266  0.07550636 0.05447396
#> factor(zyg)MZ -0.9745889 0.6188210 -1.574913 0.11527663  0.15872625 0.08263254
#> 
#> $vargam
#>             Estimate Std.Err     2.5%  97.5% P-value
#> dependence1   0.1633  0.1275 -0.08649 0.4132  0.2000
#> dependence2   0.3773  0.2335 -0.08032 0.8350  0.1061
#> 
#> $type
#> [1] "clayton.oakes"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

rm(prtsam)

dfam <- simSurvFam(1000)
dfam <- fast.reshape(dfam,var=c("x","time","status"))

desfs <- function(x,num1="num1",num2="num2")
{
pp <- (x[num1]=="m")*(x[num2]=="f")*1   ## mother-father
pc <- (x[num1]=="m" | x[num1]=="f")*(x[num2]=="b1" | x[num2]=="b2")*1 ## mother-child
cc <- (x[num1]=="b1")*(x[num2]=="b1" | x[num2]=="b2")*1               ## child-child
c(pp,pc,cc)
}

marg <- coxph(Surv(time,status)~factor(num),data=dfam)
out3 <- easy.survival.twostage(marg,data=dfam,time="time",status="status",id="id",
             deshelp=0,
             method="nr",theta.formula=desfs,
             model="plackett",
             desnames=c("parent-parent","parent-child","child-cild"))
summary(out3)

#> Dependence parameter for Odds-Ratio (Plackett) model 
#> With log-link 

#> $estimates
#>               log-Coef.         SE         z        P-val Spearman Corr.
#> parent-parent 1.4395408 0.09217263 15.617877 0.000000e+00      0.4489952
#> parent-child  0.7489145 0.06093386 12.290614 0.000000e+00      0.2450628
#> child-cild    0.6539244 0.09216016  7.095522 1.288747e-12      0.2149146
#>                       SE
#> parent-parent 0.02507264
#> parent-child  0.01920914
#> child-cild    0.02943919
#> 
#> $or
#>               Estimate Std.Err  2.5% 97.5%   P-value
#> parent-parent    4.219  0.3889 3.457 4.981 2.012e-27
#> parent-child     2.115  0.1289 1.862 2.367 1.589e-60
#> child-cild       1.923  0.1772 1.576 2.270 1.980e-27
#> 
#> $type
#> [1] "plackett"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"