Fast pattern

Fast pattern

fast.pattern(x, y, categories = 2, ...)

Arguments

x
Matrix (binary) of patterns. Optionally if y is also passed as argument, then the pattern matrix is defined as the elements agreeing in the two matrices.

y
Optional matrix argument with same dimensions as x (see above)

categories
Default 2 (binary)

...
Optional additional arguments

Examples

X <- matrix(rbinom(100,1,0.5),ncol=4)
fast.pattern(X)

#> $pattern
#>       [,1] [,2] [,3] [,4]
#>  [1,]    0    1    1    1
#>  [2,]    1    0    0    0
#>  [3,]    1    0    1    0
#>  [4,]    0    0    1    0
#>  [5,]    1    1    0    1
#>  [6,]    0    0    0    1
#>  [7,]    1    0    0    1
#>  [8,]    0    1    1    0
#>  [9,]    1    1    1    1
#> [10,]    0    0    1    1
#> [11,]    1    1    0    0
#> [12,]    0    1    0    0
#> [13,]    0    1    0    1
#> [14,]    1    0    1    1
#> 
#> $group
#>       [,1]
#>  [1,]    0
#>  [2,]    1
#>  [3,]    2
#>  [4,]    3
#>  [5,]    4
#>  [6,]    5
#>  [7,]    6
#>  [8,]    7
#>  [9,]    1
#> [10,]    8
#> [11,]    9
#> [12,]    6
#> [13,]   10
#> [14,]   10
#> [15,]    7
#> [16,]   11
#> [17,]    3
#> [18,]    8
#> [19,]   12
#> [20,]   11
#> [21,]    9
#> [22,]    5
#> [23,]   13
#> [24,]    4
#> [25,]   11
#> 

X <- matrix(rbinom(100,3,0.5),ncol=4)
fast.pattern(X,categories=4)

#> $pattern
#>       [,1] [,2] [,3] [,4]
#>  [1,]    1    2    1    2
#>  [2,]    0    3    2    0
#>  [3,]    2    2    1    1
#>  [4,]    0    3    2    3
#>  [5,]    0    1    0    2
#>  [6,]    2    0    1    2
#>  [7,]    2    1    2    1
#>  [8,]    3    1    3    2
#>  [9,]    0    2    1    1
#> [10,]    2    2    2    2
#> [11,]    1    2    0    0
#> [12,]    1    2    2    0
#> [13,]    2    2    1    3
#> [14,]    1    3    2    2
#> [15,]    1    1    0    3
#> [16,]    2    0    1    1
#> [17,]    1    0    1    1
#> [18,]    3    2    2    2
#> [19,]    2    1    2    2
#> [20,]    1    1    2    1
#> [21,]    2    1    0    1
#> [22,]    3    2    0    3
#> 
#> $group
#>       [,1]
#>  [1,]    0
#>  [2,]    1
#>  [3,]    2
#>  [4,]    3
#>  [5,]    4
#>  [6,]    5
#>  [7,]    6
#>  [8,]    7
#>  [9,]    8
#> [10,]    9
#> [11,]   10
#> [12,]   11
#> [13,]   12
#> [14,]   13
#> [15,]   14
#> [16,]   15
#> [17,]   16
#> [18,]   17
#> [19,]   13
#> [20,]    6
#> [21,]    6
#> [22,]   18
#> [23,]   19
#> [24,]   20
#> [25,]   21
#>