Fast reshape

Fast reshape/tranpose of data

fast.reshape(
  data,
  varying,
  id,
  num,
  sep = "",
  keep,
  idname = "id",
  numname = "num",
  factor = FALSE,
  idcombine = TRUE,
  labelnum = FALSE,
  labels,
  regex = mets.options()$regex,
  dropid = FALSE,
  ...
)

Arguments

data
data.frame or matrix

varying
Vector of prefix-names of the time varying variables. Optional for Long->Wide reshaping.

id
id-variable. If omitted then reshape Wide->Long.

num
Optional number/time variable

sep
String seperating prefix-name with number/time

keep
Vector of column names to keep

idname
Name of id-variable (Wide->Long)

numname
Name of number-variable (Wide->Long)

factor
If true all factors are kept (otherwise treated as character)

idcombine
If TRUE and id is vector of several variables, the unique id is combined from all the variables. Otherwise the first variable is only used as identifier.

labelnum
If TRUE varying variables in wide format (going from long->wide) are labeled 1,2,3,... otherwise use 'num' variable. In long-format (going from wide->long) varying variables matching 'varying' prefix are only selected if their postfix is a number.

labels
Optional labels for the number variable

regex
Use regular expressions

dropid
Drop id in long format (default FALSE)

...
Optional additional arguments

Author

Thomas Scheike, Klaus K. Holst

Examples

library("lava")
m <- lvm(c(y1,y2,y3,y4)~x)
d <- sim(m,5)
d

#>           y1         y2        y3         y4          x
#> 1  2.2363835  0.4946415  1.230584  0.3874222  0.9621697
#> 2  2.2162689  1.7399893  2.007292  2.4714479  1.3152824
#> 3 -0.9677689 -1.8707039 -0.905182 -2.2929790 -1.9867900
#> 4 -0.7195039 -2.5031543 -2.188153 -1.2104840 -0.4562290
#> 5 -0.7458861  0.5135096 -2.062469 -0.1541229 -0.2447923

fast.reshape(d,"y")

#>             x          y id num
#> 1   0.9621697  2.2363835  1   1
#> 2   0.9621697  0.4946415  1   2
#> 3   0.9621697  1.2305837  1   3
#> 4   0.9621697  0.3874222  1   4
#> 5   1.3152824  2.2162689  2   1
#> 6   1.3152824  1.7399893  2   2
#> 7   1.3152824  2.0072921  2   3
#> 8   1.3152824  2.4714479  2   4
#> 9  -1.9867900 -0.9677689  3   1
#> 10 -1.9867900 -1.8707039  3   2
#> 11 -1.9867900 -0.9051820  3   3
#> 12 -1.9867900 -2.2929790  3   4
#> 13 -0.4562290 -0.7195039  4   1
#> 14 -0.4562290 -2.5031543  4   2
#> 15 -0.4562290 -2.1881526  4   3
#> 16 -0.4562290 -1.2104840  4   4
#> 17 -0.2447923 -0.7458861  5   1
#> 18 -0.2447923  0.5135096  5   2
#> 19 -0.2447923 -2.0624691  5   3
#> 20 -0.2447923 -0.1541229  5   4

fast.reshape(fast.reshape(d,"y"),id="id")

#>           x1         y1 id num1         x2         y2 num2         x3        y3
#> 1  0.9621697  2.2363835  1    1  0.9621697  0.4946415    2  0.9621697  1.230584
#> 2  1.3152824  2.2162689  2    1  1.3152824  1.7399893    2  1.3152824  2.007292
#> 3 -1.9867900 -0.9677689  3    1 -1.9867900 -1.8707039    2 -1.9867900 -0.905182
#> 4 -0.4562290 -0.7195039  4    1 -0.4562290 -2.5031543    2 -0.4562290 -2.188153
#> 5 -0.2447923 -0.7458861  5    1 -0.2447923  0.5135096    2 -0.2447923 -2.062469
#>   num3         x4         y4 num4
#> 1    3  0.9621697  0.3874222    4
#> 2    3  1.3152824  2.4714479    4
#> 3    3 -1.9867900 -2.2929790    4
#> 4    3 -0.4562290 -1.2104840    4
#> 5    3 -0.2447923 -0.1541229    4

##### From wide-format
(dd <- fast.reshape(d,"y"))

#>             x          y id num
#> 1   0.9621697  2.2363835  1   1
#> 2   0.9621697  0.4946415  1   2
#> 3   0.9621697  1.2305837  1   3
#> 4   0.9621697  0.3874222  1   4
#> 5   1.3152824  2.2162689  2   1
#> 6   1.3152824  1.7399893  2   2
#> 7   1.3152824  2.0072921  2   3
#> 8   1.3152824  2.4714479  2   4
#> 9  -1.9867900 -0.9677689  3   1
#> 10 -1.9867900 -1.8707039  3   2
#> 11 -1.9867900 -0.9051820  3   3
#> 12 -1.9867900 -2.2929790  3   4
#> 13 -0.4562290 -0.7195039  4   1
#> 14 -0.4562290 -2.5031543  4   2
#> 15 -0.4562290 -2.1881526  4   3
#> 16 -0.4562290 -1.2104840  4   4
#> 17 -0.2447923 -0.7458861  5   1
#> 18 -0.2447923  0.5135096  5   2
#> 19 -0.2447923 -2.0624691  5   3
#> 20 -0.2447923 -0.1541229  5   4

## Same with explicit setting new id and number variable/column names
## and seperator "" (default) and dropping x
fast.reshape(d,"y",idname="a",timevar="b",sep="",keep=c())

#>             y a num
#> 1   2.2363835 1   1
#> 2   0.4946415 1   2
#> 3   1.2305837 1   3
#> 4   0.3874222 1   4
#> 5   2.2162689 2   1
#> 6   1.7399893 2   2
#> 7   2.0072921 2   3
#> 8   2.4714479 2   4
#> 9  -0.9677689 3   1
#> 10 -1.8707039 3   2
#> 11 -0.9051820 3   3
#> 12 -2.2929790 3   4
#> 13 -0.7195039 4   1
#> 14 -2.5031543 4   2
#> 15 -2.1881526 4   3
#> 16 -1.2104840 4   4
#> 17 -0.7458861 5   1
#> 18  0.5135096 5   2
#> 19 -2.0624691 5   3
#> 20 -0.1541229 5   4

## Same with 'reshape' list-syntax
fast.reshape(d,list(c("y1","y2","y3","y4")),labelnum=TRUE)

#>             x         y1 id num
#> 1   0.9621697  2.2363835  1   1
#> 2   0.9621697  0.4946415  1   2
#> 3   0.9621697  1.2305837  1   3
#> 4   0.9621697  0.3874222  1   4
#> 5   1.3152824  2.2162689  2   1
#> 6   1.3152824  1.7399893  2   2
#> 7   1.3152824  2.0072921  2   3
#> 8   1.3152824  2.4714479  2   4
#> 9  -1.9867900 -0.9677689  3   1
#> 10 -1.9867900 -1.8707039  3   2
#> 11 -1.9867900 -0.9051820  3   3
#> 12 -1.9867900 -2.2929790  3   4
#> 13 -0.4562290 -0.7195039  4   1
#> 14 -0.4562290 -2.5031543  4   2
#> 15 -0.4562290 -2.1881526  4   3
#> 16 -0.4562290 -1.2104840  4   4
#> 17 -0.2447923 -0.7458861  5   1
#> 18 -0.2447923  0.5135096  5   2
#> 19 -0.2447923 -2.0624691  5   3
#> 20 -0.2447923 -0.1541229  5   4

##### From long-format
fast.reshape(dd,id="id")

#>           x1         y1 id num1         x2         y2 num2         x3        y3
#> 1  0.9621697  2.2363835  1    1  0.9621697  0.4946415    2  0.9621697  1.230584
#> 2  1.3152824  2.2162689  2    1  1.3152824  1.7399893    2  1.3152824  2.007292
#> 3 -1.9867900 -0.9677689  3    1 -1.9867900 -1.8707039    2 -1.9867900 -0.905182
#> 4 -0.4562290 -0.7195039  4    1 -0.4562290 -2.5031543    2 -0.4562290 -2.188153
#> 5 -0.2447923 -0.7458861  5    1 -0.2447923  0.5135096    2 -0.2447923 -2.062469
#>   num3         x4         y4 num4
#> 1    3  0.9621697  0.3874222    4
#> 2    3  1.3152824  2.4714479    4
#> 3    3 -1.9867900 -2.2929790    4
#> 4    3 -0.4562290 -1.2104840    4
#> 5    3 -0.2447923 -0.1541229    4

## Restrict set up within-cluster varying variables
fast.reshape(dd,"y",id="id")

#>            x         y1 id num         y2        y3         y4
#> 1  0.9621697  2.2363835  1   1  0.4946415  1.230584  0.3874222
#> 2  1.3152824  2.2162689  2   1  1.7399893  2.007292  2.4714479
#> 3 -1.9867900 -0.9677689  3   1 -1.8707039 -0.905182 -2.2929790
#> 4 -0.4562290 -0.7195039  4   1 -2.5031543 -2.188153 -1.2104840
#> 5 -0.2447923 -0.7458861  5   1  0.5135096 -2.062469 -0.1541229

fast.reshape(dd,"y",id="id",keep="x",sep=".")

#>            x        y.1 id        y.2       y.3        y.4
#> 1  0.9621697  2.2363835  1  0.4946415  1.230584  0.3874222
#> 2  1.3152824  2.2162689  2  1.7399893  2.007292  2.4714479
#> 3 -1.9867900 -0.9677689  3 -1.8707039 -0.905182 -2.2929790
#> 4 -0.4562290 -0.7195039  4 -2.5031543 -2.188153 -1.2104840
#> 5 -0.2447923 -0.7458861  5  0.5135096 -2.062469 -0.1541229

#####
x <- data.frame(id=c(5,5,6,6,7),y=1:5,x=1:5,tv=c(1,2,2,1,2))
x

#>   id y x tv
#> 1  5 1 1  1
#> 2  5 2 2  2
#> 3  6 3 3  2
#> 4  6 4 4  1
#> 5  7 5 5  2

(xw <- fast.reshape(x,id="id"))

#>   id y1 x1 tv1 y2 x2 tv2
#> 1  5  1  1   1  2  2   2
#> 2  6  3  3   2  4  4   1
#> 3  7  5  5   2 NA NA  NA

(xl <- fast.reshape(xw,c("y","x"),idname="id2",keep=c()))

#>    y  x id2 num
#> 1  1  1   1   1
#> 2  2  2   1   2
#> 3  3  3   2   1
#> 4  4  4   2   2
#> 5  5  5   3   1
#> 6 NA NA   3   2

(xl <- fast.reshape(xw,c("y","x","tv")))

#>   id  y  x tv num
#> 1  5  1  1  1   1
#> 2  5  2  2  2   2
#> 3  6  3  3  2   1
#> 4  6  4  4  1   2
#> 5  7  5  5  2   1
#> 6  7 NA NA NA   2

(xw2 <- fast.reshape(xl,id="id",num="num"))

#>   id y1 x1 tv1 y2 x2 tv2
#> 1  5  1  1   1  2  2   2
#> 2  6  3  3   2  4  4   1
#> 3  7  5  5   2 NA NA  NA

fast.reshape(xw2,c("y","x"),idname="id")

#>   id tv1 tv2  y  x num
#> 1  5   1   2  1  1   1
#> 2  5   1   2  2  2   2
#> 3  6   2   1  3  3   1
#> 4  6   2   1  4  4   2
#> 5  7   2  NA  5  5   1
#> 6  7   2  NA NA NA   2

### more generally:
### varying=list(c("ym","yf","yb1","yb2"), c("zm","zf","zb1","zb2"))
### varying=list(c("ym","yf","yb1","yb2")))

##### Family cluster example
d <- mets:::simBinFam(3)
d

#>       agem     agef     ageb1    ageb2 xm xf xb1 xb2 ym yf yb1 yb2 id
#> 1 26.09384 27.36191  9.096248 10.66697  1  1   1   1  1  1   1   1  1
#> 2 29.41234 32.91351 14.899234 16.33437  1  1   1   1  1  1   1   1  2
#> 3 23.75109 24.91365 13.758275 15.12345  1  1   1   0  1  1   0   0  3

fast.reshape(d,var="y")

#>        agem     agef     ageb1    ageb2 xm xf xb1 xb2 id y num
#> 1  26.09384 27.36191  9.096248 10.66697  1  1   1   1  1 1   m
#> 2  26.09384 27.36191  9.096248 10.66697  1  1   1   1  1 1   f
#> 3  26.09384 27.36191  9.096248 10.66697  1  1   1   1  1 1  b1
#> 4  26.09384 27.36191  9.096248 10.66697  1  1   1   1  1 1  b2
#> 5  29.41234 32.91351 14.899234 16.33437  1  1   1   1  2 1   m
#> 6  29.41234 32.91351 14.899234 16.33437  1  1   1   1  2 1   f
#> 7  29.41234 32.91351 14.899234 16.33437  1  1   1   1  2 1  b1
#> 8  29.41234 32.91351 14.899234 16.33437  1  1   1   1  2 1  b2
#> 9  23.75109 24.91365 13.758275 15.12345  1  1   1   0  3 1   m
#> 10 23.75109 24.91365 13.758275 15.12345  1  1   1   0  3 1   f
#> 11 23.75109 24.91365 13.758275 15.12345  1  1   1   0  3 0  b1
#> 12 23.75109 24.91365 13.758275 15.12345  1  1   1   0  3 0  b2

fast.reshape(d,varying=list(c("ym","yf","yb1","yb2")))

#>        agem     agef     ageb1    ageb2 xm xf xb1 xb2 id ym num
#> 1  26.09384 27.36191  9.096248 10.66697  1  1   1   1  1  1  ym
#> 2  26.09384 27.36191  9.096248 10.66697  1  1   1   1  1  1  yf
#> 3  26.09384 27.36191  9.096248 10.66697  1  1   1   1  1  1 yb1
#> 4  26.09384 27.36191  9.096248 10.66697  1  1   1   1  1  1 yb2
#> 5  29.41234 32.91351 14.899234 16.33437  1  1   1   1  2  1  ym
#> 6  29.41234 32.91351 14.899234 16.33437  1  1   1   1  2  1  yf
#> 7  29.41234 32.91351 14.899234 16.33437  1  1   1   1  2  1 yb1
#> 8  29.41234 32.91351 14.899234 16.33437  1  1   1   1  2  1 yb2
#> 9  23.75109 24.91365 13.758275 15.12345  1  1   1   0  3  1  ym
#> 10 23.75109 24.91365 13.758275 15.12345  1  1   1   0  3  1  yf
#> 11 23.75109 24.91365 13.758275 15.12345  1  1   1   0  3  0 yb1
#> 12 23.75109 24.91365 13.758275 15.12345  1  1   1   0  3  0 yb2

d <- sim(lvm(~y1+y2+ya),10)
d

#>             y1          y2          ya
#> 1   1.33764478 -0.41504725  0.26237568
#> 2  -0.10942127  0.34119368 -0.45628311
#> 3   0.91679581  1.39342931 -0.60957449
#> 4   0.98043076 -0.60456361 -3.80766246
#> 5  -1.00970774  0.61192812  1.31799454
#> 6   0.32508586  1.08371191 -0.54277113
#> 7  -0.14178005  0.52109427  0.46564005
#> 8  -2.22750533  0.06642853 -0.18651948
#> 9  -1.02616727  0.51126046 -0.04165241
#> 10 -0.05472261 -1.59965484 -0.34129512

(dd <- fast.reshape(d,"y"))

#>              y id num
#> 1   1.33764478  1   1
#> 2  -0.41504725  1   2
#> 3   0.26237568  1   a
#> 4  -0.10942127  2   1
#> 5   0.34119368  2   2
#> 6  -0.45628311  2   a
#> 7   0.91679581  3   1
#> 8   1.39342931  3   2
#> 9  -0.60957449  3   a
#> 10  0.98043076  4   1
#> 11 -0.60456361  4   2
#> 12 -3.80766246  4   a
#> 13 -1.00970774  5   1
#> 14  0.61192812  5   2
#> 15  1.31799454  5   a
#> 16  0.32508586  6   1
#> 17  1.08371191  6   2
#> 18 -0.54277113  6   a
#> 19 -0.14178005  7   1
#> 20  0.52109427  7   2
#> 21  0.46564005  7   a
#> 22 -2.22750533  8   1
#> 23  0.06642853  8   2
#> 24 -0.18651948  8   a
#> 25 -1.02616727  9   1
#> 26  0.51126046  9   2
#> 27 -0.04165241  9   a
#> 28 -0.05472261 10   1
#> 29 -1.59965484 10   2
#> 30 -0.34129512 10   a

fast.reshape(d,"y",labelnum=TRUE)

#>             ya           y id num
#> 1   0.26237568  1.33764478  1   1
#> 2   0.26237568 -0.41504725  1   2
#> 3  -0.45628311 -0.10942127  2   1
#> 4  -0.45628311  0.34119368  2   2
#> 5  -0.60957449  0.91679581  3   1
#> 6  -0.60957449  1.39342931  3   2
#> 7  -3.80766246  0.98043076  4   1
#> 8  -3.80766246 -0.60456361  4   2
#> 9   1.31799454 -1.00970774  5   1
#> 10  1.31799454  0.61192812  5   2
#> 11 -0.54277113  0.32508586  6   1
#> 12 -0.54277113  1.08371191  6   2
#> 13  0.46564005 -0.14178005  7   1
#> 14  0.46564005  0.52109427  7   2
#> 15 -0.18651948 -2.22750533  8   1
#> 16 -0.18651948  0.06642853  8   2
#> 17 -0.04165241 -1.02616727  9   1
#> 18 -0.04165241  0.51126046  9   2
#> 19 -0.34129512 -0.05472261 10   1
#> 20 -0.34129512 -1.59965484 10   2

fast.reshape(dd,id="id",num="num")

#>             y1 id          y2          ya
#> 1   1.33764478  1 -0.41504725  0.26237568
#> 2  -0.10942127  2  0.34119368 -0.45628311
#> 3   0.91679581  3  1.39342931 -0.60957449
#> 4   0.98043076  4 -0.60456361 -3.80766246
#> 5  -1.00970774  5  0.61192812  1.31799454
#> 6   0.32508586  6  1.08371191 -0.54277113
#> 7  -0.14178005  7  0.52109427  0.46564005
#> 8  -2.22750533  8  0.06642853 -0.18651948
#> 9  -1.02616727  9  0.51126046 -0.04165241
#> 10 -0.05472261 10 -1.59965484 -0.34129512

fast.reshape(dd,id="id",num="num",labelnum=TRUE)

#>             y1 id          y2          y3
#> 1   1.33764478  1 -0.41504725  0.26237568
#> 2  -0.10942127  2  0.34119368 -0.45628311
#> 3   0.91679581  3  1.39342931 -0.60957449
#> 4   0.98043076  4 -0.60456361 -3.80766246
#> 5  -1.00970774  5  0.61192812  1.31799454
#> 6   0.32508586  6  1.08371191 -0.54277113
#> 7  -0.14178005  7  0.52109427  0.46564005
#> 8  -2.22750533  8  0.06642853 -0.18651948
#> 9  -1.02616727  9  0.51126046 -0.04165241
#> 10 -0.05472261 10 -1.59965484 -0.34129512

fast.reshape(d,c(a="y"),labelnum=TRUE) ## New column name

#>             ya           a id num
#> 1   0.26237568  1.33764478  1   1
#> 2   0.26237568 -0.41504725  1   2
#> 3  -0.45628311 -0.10942127  2   1
#> 4  -0.45628311  0.34119368  2   2
#> 5  -0.60957449  0.91679581  3   1
#> 6  -0.60957449  1.39342931  3   2
#> 7  -3.80766246  0.98043076  4   1
#> 8  -3.80766246 -0.60456361  4   2
#> 9   1.31799454 -1.00970774  5   1
#> 10  1.31799454  0.61192812  5   2
#> 11 -0.54277113  0.32508586  6   1
#> 12 -0.54277113  1.08371191  6   2
#> 13  0.46564005 -0.14178005  7   1
#> 14  0.46564005  0.52109427  7   2
#> 15 -0.18651948 -2.22750533  8   1
#> 16 -0.18651948  0.06642853  8   2
#> 17 -0.04165241 -1.02616727  9   1
#> 18 -0.04165241  0.51126046  9   2
#> 19 -0.34129512 -0.05472261 10   1
#> 20 -0.34129512 -1.59965484 10   2

##### Unbalanced data
m <- lvm(c(y1,y2,y3,y4)~ x+z1+z3+z5)
d <- sim(m,3)
d

#>           y1         y2        y3         y4          x          z1          z3
#> 1  2.0533778  1.7691243  2.590250  2.4621101 -0.2905331  0.65021574  0.05225591
#> 2 -1.6883598 -2.2693025 -3.258077 -0.6541728  1.5555723 -1.68832368 -0.48894459
#> 3  0.5963096  0.7015472  1.569971 -1.7878806  0.7315199 -0.05380033 -1.20277031
#>           z5
#> 1  0.9821673
#> 2 -1.1820431
#> 3  0.1141341

fast.reshape(d,c("y","z"))

#>             x          y           z id num
#> 1  -0.2905331  2.0533778  0.65021574  1   1
#> 2  -0.2905331  1.7691243          NA  1   2
#> 3  -0.2905331  2.5902502  0.05225591  1   3
#> 4  -0.2905331  2.4621101          NA  1   4
#> 5  -0.2905331         NA  0.98216728  1   5
#> 6   1.5555723 -1.6883598 -1.68832368  2   1
#> 7   1.5555723 -2.2693025          NA  2   2
#> 8   1.5555723 -3.2580765 -0.48894459  2   3
#> 9   1.5555723 -0.6541728          NA  2   4
#> 10  1.5555723         NA -1.18204306  2   5
#> 11  0.7315199  0.5963096 -0.05380033  3   1
#> 12  0.7315199  0.7015472          NA  3   2
#> 13  0.7315199  1.5699708 -1.20277031  3   3
#> 14  0.7315199 -1.7878806          NA  3   4
#> 15  0.7315199         NA  0.11413410  3   5

##### not-varying syntax:
fast.reshape(d,-c("x"))

#>             x          y           z id num
#> 1  -0.2905331  2.0533778  0.65021574  1   1
#> 2  -0.2905331  1.7691243          NA  1   2
#> 3  -0.2905331  2.5902502  0.05225591  1   3
#> 4  -0.2905331  2.4621101          NA  1   4
#> 5  -0.2905331         NA  0.98216728  1   5
#> 6   1.5555723 -1.6883598 -1.68832368  2   1
#> 7   1.5555723 -2.2693025          NA  2   2
#> 8   1.5555723 -3.2580765 -0.48894459  2   3
#> 9   1.5555723 -0.6541728          NA  2   4
#> 10  1.5555723         NA -1.18204306  2   5
#> 11  0.7315199  0.5963096 -0.05380033  3   1
#> 12  0.7315199  0.7015472          NA  3   2
#> 13  0.7315199  1.5699708 -1.20277031  3   3
#> 14  0.7315199 -1.7878806          NA  3   4
#> 15  0.7315199         NA  0.11413410  3   5

##### Automatically define varying variables from trailing digits
fast.reshape(d)

#>             x          y           z id num
#> 1  -0.2905331  2.0533778  0.65021574  1   1
#> 2  -0.2905331  1.7691243          NA  1   2
#> 3  -0.2905331  2.5902502  0.05225591  1   3
#> 4  -0.2905331  2.4621101          NA  1   4
#> 5  -0.2905331         NA  0.98216728  1   5
#> 6   1.5555723 -1.6883598 -1.68832368  2   1
#> 7   1.5555723 -2.2693025          NA  2   2
#> 8   1.5555723 -3.2580765 -0.48894459  2   3
#> 9   1.5555723 -0.6541728          NA  2   4
#> 10  1.5555723         NA -1.18204306  2   5
#> 11  0.7315199  0.5963096 -0.05380033  3   1
#> 12  0.7315199  0.7015472          NA  3   2
#> 13  0.7315199  1.5699708 -1.20277031  3   3
#> 14  0.7315199 -1.7878806          NA  3   4
#> 15  0.7315199         NA  0.11413410  3   5

##### Prostate cancer example
data(prt)
head(prtw <- fast.reshape(prt,"cancer",id="id"))

#>    country      time status zyg id cancer1 cancer2
#> 31 Denmark  96.98833      1  DZ  1       0       0
#> 39 Denmark  68.04498      1  DZ  3       0       0
#> 51 Denmark  78.78068      1  DZ  5       0       0
#> 70 Denmark 100.95488      1  MZ  9       0       0
#> 83 Denmark 104.55035      1  DZ 12       0       1
#> 95 Denmark  95.65324      1  DZ 15       0       0

ftable(cancer1~cancer2,data=prtw)

#>         cancer1     0     1
#> cancer2                    
#> 0               13405   349
#> 1                 362   106

rm(prtw)