Lu-Tsiatis More Efficient Log-Rank for Randomized studies with baseline covariates

Efficient implementation of the Lu-Tsiatis improvement using baseline covariates, extended to competing risks and recurrent events. Results almost equivalent with the speffSurv function of the speff2trial function in the survival case. A dynamic censoring augmentation regression is also computed to gain even more from the censoring augmentation. Furhter, we also deal with twostage randomizations. The function was implemented to deal with recurrent events (start,stop) + cluster, and more examples in vignette.

phreg_rct(
  formula,
  data,
  cause = 1,
  cens.code = 0,
  typesR = c("R0", "R1", "R01"),
  typesC = c("C", "dynC"),
  augmentR0 = NULL,
  augmentR1 = NULL,
  augmentC = NULL,
  treat.model = ~+1,
  RCT = TRUE,
  weight.var = NULL,
  km = TRUE,
  level = 0.95,
  cens.model = NULL,
  estpr = 1,
  pi0 = 0.5,
  ...
)

Arguments

formula

formula with 'Surv' or 'Event' outcome (see coxph) and treatment (randomization 0/1)

data

data frame

cause

to use for competing risks, recurrent events data

cens.code

to use for competing risks, recurrent events data

typesR

augmentations used for randomization

typesC

augmentations used for censoring

augmentR0

formula for the randomization augmentation (~age+sex)

augmentR1

formula for the randomization augmentation (~age+sex)

augmentC

formula for the censoring augmentation (~age+sex)

treat.model

propensity score model, default is ~+1, assuming RCT study

RCT

if false will use propensity score adjustment

weight.var

in case of twostage randomization, this variable is 1 for the treatment times,

km

use Kaplan-Meier for the censoring weights (stratified on treatment)

level

of confidence intervals

cens.model,

default is censoring model ~strata(treatment) but any model can be used to make censoring martingales

estpr

estimates propensity scores

pi0

possible fixed propoensity scores for randomizations

...

Additional arguments to phreg function

References

Lu, Tsiatis (2008), Improving the efficiency of the log-rank test using auxiliary covariates, Biometrika, 679--694 Scheike (2024), WIP, Two-stage randomization for recurrent events,

Author

Thomas Scheike

Examples

## Lu, Tsiatis simulation
data <- mets:::simLT(0.7,100)
dfactor(data) <- Z.f~Z

out <- phreg_rct(Surv(time,status)~Z.f,data=data,augmentR0=~X,augmentC=~factor(Z):X)
summary(out)
#>                 Estimate   Std.Err       2.5%     97.5%   P-value
#> Marginal-Z.f1 0.17228267 0.2727585 -0.3623143 0.7068796 0.5276282
#> R0_C:Z.f1     0.10449281 0.2236195 -0.3337933 0.5427790 0.6402999
#> R0_dynC:Z.f1  0.02207617 0.2193279 -0.4077986 0.4519509 0.9198253
#> attr(,"class")
#> [1] "summary.phreg_rct"