Average Treatment effect for Restricted Mean for censored competing risks data using IPCW

Under the standard causal assumptions we can estimate the average treatment effect E(Y(1) - Y(0)). We need Consistency, ignorability ( Y(1), Y(0) indep A given X), and positivity.

resmeanATE(formula, data, model = "exp", ...)

Arguments

formula

formula with 'Event' outcome

data

data-frame

model

possible exp model for relevant mean model that is exp(X^t beta)

...

Additional arguments to pass to binregATE

Details

The first covariate in the specification of the competing risks regression model must be the treatment effect that is a factor. If the factor has more than two levels then it uses the mlogit for propensity score modelling. We consider the outcome mint(T;tau) or I(epsion==cause1)(t- min(T;t)) that gives years lost due to cause "cause" depending on the number of causes. The default model is the exp(X^ beta) and otherwise a linear model is used.

Estimates the ATE using the the standard binary double robust estimating equations that are IPCW censoring adjusted.

Author

Thomas Scheike

Examples

library(mets); data(bmt); bmt$event <- bmt$cause!=0; dfactor(bmt) <- tcell~tcell
out <- resmeanATE(Event(time,event)~tcell+platelet,data=bmt,time=40,treat.model=tcell~platelet)
summary(out)
#> 
#>    n events
#>  408    241
#> 
#>  408 clusters
#> coeffients:
#>              Estimate   Std.Err      2.5%     97.5% P-value
#> (Intercept)  2.852563  0.062496  2.730074  2.975052  0.0000
#> tcell1       0.021289  0.122983 -0.219753  0.262330  0.8626
#> platelet     0.303313  0.090772  0.125404  0.481222  0.0008
#> 
#> exp(coeffients):
#>             Estimate     2.5%   97.5%
#> (Intercept) 17.33214 15.33402 19.5906
#> tcell1       1.02152  0.80272  1.3000
#> platelet     1.35434  1.13361  1.6181
#> 
#> Average Treatment effects (G-formula) :
#>           Estimate  Std.Err     2.5%    97.5% P-value
#> treat0    19.25887  0.95918 17.37892 21.13882  0.0000
#> treat1    19.67326  2.22868 15.30513 24.04139  0.0000
#> treat:1-0  0.41439  2.41150 -4.31207  5.14085  0.8636
#> 
#> Average Treatment effects (double robust) :
#>           Estimate Std.Err    2.5%   97.5% P-value
#> treat0     19.3224  1.0515 17.2614 21.3834  0.0000
#> treat1     21.5582  3.8016 14.1072 29.0091  0.0000
#> treat:1-0   2.2358  4.1989 -5.9940 10.4655  0.5944
#> 
#> 

out1 <- resmeanATE(Event(time,cause)~tcell+platelet,data=bmt,cause=1,time=40,
                   treat.model=tcell~platelet)
summary(out1)
#> 
#>    n events
#>  408    157
#> 
#>  408 clusters
#> coeffients:
#>              Estimate   Std.Err      2.5%     97.5% P-value
#> (Intercept)  2.806269  0.069619  2.669817  2.942721  0.0000
#> tcell1      -0.374141  0.247689 -0.859602  0.111320  0.1309
#> platelet    -0.491646  0.164934 -0.814911 -0.168381  0.0029
#> 
#> exp(coeffients):
#>             Estimate     2.5%   97.5%
#> (Intercept) 16.54807 14.43733 18.9674
#> tcell1       0.68788  0.42333  1.1178
#> platelet     0.61162  0.44268  0.8450
#> 
#> Average Treatment effects (G-formula) :
#>           Estimate  Std.Err     2.5%    97.5% P-value
#> treat0    14.53176  0.95705 12.65599 16.40754  0.0000
#> treat1     9.99611  2.37816  5.33501 14.65721  0.0000
#> treat:1-0 -4.53566  2.57515 -9.58286  0.51155  0.0782
#> 
#> Average Treatment effects (double robust) :
#>              Estimate     Std.Err        2.5%       97.5% P-value
#> treat0     14.5200247   0.9576900  12.6429868  16.3970625  0.0000
#> treat1      9.4568371   2.4051437   4.7428421  14.1708321  0.0001
#> treat:1-0  -5.0631876   2.5856762 -10.1310197   0.0046446  0.0502
#> 
#>