Restricted IPCW mean for censored survival data

Simple version of comp.risk function of timereg for just one time-point thus fitting the model $$E(T \leq t | X ) = expit( X^T beta) $$

resmeanIPCW(
  formula,
  data,
  cause = 1,
  time = NULL,
  beta = NULL,
  offset = NULL,
  weights = NULL,
  cens.weights = NULL,
  cens.model = ~+1,
  se = TRUE,
  kaplan.meier = TRUE,
  cens.code = 0,
  no.opt = FALSE,
  method = "nr",
  model = "exp",
  augmentation = NULL,
  h = NULL,
  MCaugment = NULL,
  ...
)

Arguments

formula

formula with outcome (see coxph)

data

data frame

cause

cause of interest

time

time of interest

beta

starting values

offset

offsets for partial likelihood

weights

for score equations

cens.weights

censoring weights

cens.model

only stratified cox model without covariates

se

to compute se's based on IPCW

kaplan.meier

uses Kaplan-Meier for IPCW in contrast to exp(-Baseline)

cens.code

gives censoring code

no.opt

to not optimize

method

for optimization

model

exp or linear

augmentation

to augment binomial regression

h

h for estimating equation

MCaugment

iid of h and censoring model

...

Additional arguments to lower level funtions

Details

Based on binomial regresion IPCW response estimating equation: $$ X ( \Delta (T \leq t)/G_c(T_i-) - expit( X^T beta)) = 0 $$ for IPCW adjusted responses.

Based on binomial regresion IPCW response estimating equation: $$ h(X) X ( \Delta (T \leq t)/G_c(T_i-) - expit( X^T beta)) = 0 $$ for IPCW adjusted responses where $h$ is given as an argument together with iid of censoring with h.

By using appropriately the h argument we can also do the efficient IPCW estimator estimator.

Variance is based on $$ \sum w_i^2 $$ also with IPCW adjustment, and naive.var is variance under known censoring model.

Censoring model may depend on strata.

Author

Thomas Scheike

Examples

data(bmt); bmt$time <- bmt$time+runif(nrow(bmt))*0.001
# logistic regresion with IPCW binomial regression 
out <- resmeanIPCW(Event(time,cause!=0)~tcell+platelet+age,bmt,
                cause=1,time=50,cens.model=~strata(platelet),model="exp")
summary(out)
#> 
#>    n events
#>  408    245
#> 
#>  408 clusters
#> coeffients:
#>              Estimate   Std.Err      2.5%     97.5% P-value
#> (Intercept)  3.041609  0.069813  2.904778  3.178441  0.0000
#> tcell       -0.155907  0.272709 -0.690406  0.378592  0.5675
#> platelet     0.244590  0.107741  0.033422  0.455758  0.0232
#> age         -0.209240  0.085880 -0.377563 -0.040918  0.0148
#> 
#> exp(coeffients):
#>             Estimate     2.5%   97.5%
#> (Intercept) 20.93891 18.26118 24.0093
#> tcell        0.85564  0.50137  1.4602
#> platelet     1.27710  1.03399  1.5774
#> age          0.81120  0.68553  0.9599
#> 
#> 

### same as Kaplan-Meier for full censoring model 
bmt$int <- with(bmt,interaction(tcell,platelet))
out <- resmeanIPCW(Event(time,cause!=0)~-1+int,bmt,cause=1,time=30,
                            cens.model=~strata(platelet,tcell),model="lin")
summary(out)
#> 
#>    n events
#>  408    231
#> 
#>  408 clusters
#> coeffients:
#>        Estimate  Std.Err     2.5%    97.5% P-value
#> int0.0 13.60293  0.83159 11.97305 15.23281       0
#> int1.0 16.19115  2.40608 11.47532 20.90698       0
#> int0.1 18.90125  1.26964 16.41280 21.38971       0
#> int1.1 17.76614  2.45363 12.95711 22.57517       0
#> 
#> exp(coeffients):
#>         Estimate      2.5%      97.5%
#> int0.0    808494    158427    4125961
#> int1.0  10757886     96309 1201674997
#> int0.1 161700032  13427257 1947300163
#> int1.1  51968193    423841 6371953477
#> 
#> 
out1 <- phreg(Surv(time,cause!=0)~strata(tcell,platelet),data=bmt)
rm1 <- resmean.phreg(out1,times=30)
summary(rm1)
#>   strata times    rmean  se.rmean
#> 1      0    30 13.60293 0.8300096
#> 2      1    30 18.90125 1.2608700
#> 3      2    30 16.19115 2.3440336
#> 4      3    30 17.76614 2.3905379