Twostage survival model for multivariate survival data

Fits Clayton-Oakes or bivariate Plackett models for bivariate survival data using marginals that are on Cox form. The dependence can be modelled via

Regression design on dependence parameter.

Random effects, additive gamma model.

If clusters contain more than two subjects, we use a composite likelihood based on the pairwise bivariate models, for full MLE see twostageMLE.

The two-stage model is constructed such that given the gamma distributed random effects it is assumed that the survival functions are indpendent, and that the marginal survival functions are on Cox form (or additive form) $$ P(T > t| x) = S(t|x)= exp( -exp(x^T \beta) A_0(t) ) $$

One possibility is to model the variance within clusters via a regression design, and then one can specify a regression structure for the independent gamma distributed random effect for each cluster, such that the variance is given by $$ \theta = h( z_j^T \alpha) $$ where \(z\) is specified by theta.des, and a possible link function var.link=1 will will use the exponential link \(h(x)=exp(x)\), and var.link=0 the identity link \(h(x)=x\). The reported standard errors are based on the estimated information from the likelihood assuming that the marginals are known (unlike the twostageMLE and for the additive gamma model below).

Can also fit a structured additive gamma random effects model, such as the ACE, ADE model for survival data. In this case the random.design specificies the random effects for each subject within a cluster. This is a matrix of 1's and 0's with dimension n x d. With d random effects. For a cluster with two subjects, we let the random.design rows be \(v_1\) and \(v_2\). Such that the random effects for subject 1 is $$v_1^T (Z_1,...,Z_d)$$, for d random effects. Each random effect has an associated parameter \((\lambda_1,...,\lambda_d)\). By construction subjects 1's random effect are Gamma distributed with mean \(\lambda_j/v_1^T \lambda\) and variance \(\lambda_j/(v_1^T \lambda)^2\). Note that the random effect \(v_1^T (Z_1,...,Z_d)\) has mean 1 and variance \(1/(v_1^T \lambda)\). It is here asssumed that \(lamtot=v_1^T \lambda\) is fixed within clusters as it would be for the ACE model below.

Based on these parameters the relative contribution (the heritability, h) is equivalent to the expected values of the random effects: \(\lambda_j/v_1^T \lambda\)

The DEFAULT parametrization (var.par=1) uses the variances of the random effecs $$ \theta_j = \lambda_j/(v_1^T \lambda)^2 $$ For alternative parametrizations one can specify how the parameters relate to \(\lambda_j\) with the argument var.par=0.

For both types of models the basic model assumptions are that given the random effects of the clusters the survival distributions within a cluster are independent and ' on the form $$ P(T > t| x,z) = exp( -Z \cdot Laplace^{-1}(lamtot,lamtot,S(t|x)) ) $$ with the inverse laplace of the gamma distribution with mean 1 and variance 1/lamtot.

The parameters \((\lambda_1,...,\lambda_d)\) are related to the parameters of the model by a regression construction \(pard\) (d x k), that links the \(d\) \(\lambda\) parameters with the (k) underlying \(\theta\) parameters $$ \lambda = theta.des \theta $$ here using theta.des to specify these low-dimension association. Default is a diagonal matrix. This can be used to make structural assumptions about the variances of the random-effects as is needed for the ACE model for example.

The case.control option that can be used with the pair specification of the pairwise parts of the estimating equations. Here it is assumed that the second subject of each pair is the proband.

survival.twostage(
  margsurv,
  data = parent.frame(),
  score.method = "nr",
  Nit = 60,
  detail = 0,
  clusters = NULL,
  silent = 1,
  weights = NULL,
  control = list(),
  theta = NULL,
  theta.des = NULL,
  var.link = 1,
  baseline.iid = 1,
  step = 0.5,
  model = "clayton.oakes",
  marginal.trunc = NULL,
  marginal.survival = NULL,
  marginal.status = NULL,
  strata = NULL,
  se.clusters = NULL,
  numDeriv = 0,
  random.design = NULL,
  pairs = NULL,
  dim.theta = NULL,
  numDeriv.method = "simple",
  additive.gamma.sum = NULL,
  var.par = 1,
  cr.models = NULL,
  case.control = 0,
  ascertained = 0,
  shut.up = 0
)

Arguments

margsurv
Marginal model

data
data frame

score.method
Scoring method "nr", "nlminb", "optimize", "nlm"

Nit
Number of iterations

detail
Detail

clusters
Cluster variable

silent
Debug information

weights
Weights

control
Optimization arguments

theta
Starting values for variance components

theta.des
design for dependence parameters, when pairs are given the indeces of the theta-design for this pair, is given in pairs as column 5

var.link
Link function for variance: exp-link.

baseline.iid
to adjust for baseline estimation, using phreg function on same data.

step
Step size

model
model

marginal.trunc
marginal left truncation probabilities

marginal.survival
optional vector of marginal survival probabilities

marginal.status
related to marginal survival probabilities

strata
strata for fitting, see example

se.clusters
for clusters for se calculation with iid

numDeriv
to get numDeriv version of second derivative, otherwise uses sum of squared score

random.design
random effect design for additive gamma model, when pairs are given the indeces of the pairs random.design rows are given as columns 3:4

pairs
matrix with rows of indeces (two-columns) for the pairs considered in the pairwise composite score, useful for case-control sampling when marginal is known.

dim.theta
dimension of the theta parameter for pairs situation.

numDeriv.method
uses simple to speed up things and second derivative not so important.

additive.gamma.sum
for two.stage=0, this is specification of the lamtot in the models via a matrix that is multiplied onto the parameters theta (dimensions=(number random effects x number of theta parameters), when null then sums all parameters.

var.par
is 1 for the default parametrization with the variances of the random effects, var.par=0 specifies that the \(\lambda_j\)'s are used as parameters.

cr.models
competing risks models for two.stage=0, should be given as a list with models for each cause

case.control
assumes case control structure for "pairs" with second column being the probands, when this options is used the twostage model is profiled out via the paired estimating equations for the survival model.

ascertained
if the pair are sampled only when there is an event. This is in contrast to case.control sampling where a proband is given. This can be combined with control probands. Pair-call of twostage is needed and second column of pairs are the first jump time with an event for ascertained pairs, or time of control proband.

shut.up
to make the program more silent in the context of iterative procedures for case-control and ascertained sampling

References

Twostage estimation of additive gamma frailty models for survival data. Scheike (2019), work in progress

Shih and Louis (1995) Inference on the association parameter in copula models for bivariate survival data, Biometrics, (1995).

Glidden (2000), A Two-Stage estimator of the dependence parameter for the Clayton Oakes model, LIDA, (2000).

Measuring early or late dependence for bivariate twin data Scheike, Holst, Hjelmborg (2015), LIDA

Estimating heritability for cause specific mortality based on twins studies Scheike, Holst, Hjelmborg (2014), LIDA

Additive Gamma frailty models for competing risks data, Biometrics (2015) Eriksson and Scheike (2015),

Examples

data(diabetes)

# Marginal Cox model  with treat as covariate
margph <- phreg(Surv(time,status)~treat+cluster(id),data=diabetes)
### Clayton-Oakes, MLE
fitco1<-twostageMLE(margph,data=diabetes,theta=1.0)
summary(fitco1)

#> Dependence parameter for Clayton-Oakes model
#> Variance of Gamma distributed random effects 

#> $estimates
#>                 Coef.        SE       z       P-val Kendall tau         SE
#> dependence1 0.9526614 0.3543033 2.68883 0.007170289    0.322645 0.08127892
#> 
#> $type
#> NULL
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

### Plackett model
mph <- phreg(Surv(time,status)~treat+cluster(id),data=diabetes)
fitp <- survival.twostage(mph,data=diabetes,theta=3.0,Nit=40,
               clusters=diabetes$id,var.link=1,model="plackett")
summary(fitp)

#> Dependence parameter for Odds-Ratio (Plackett) model 
#> With log-link 

#> $estimates
#>             log-Coef.        SE        z        P-val Spearman Corr.         SE
#> dependence1   1.14188 0.2754994 4.144764 3.401635e-05      0.3648217 0.08073474
#> 
#> $or
#>             Estimate Std.Err  2.5% 97.5%   P-value
#> dependence1    3.133   0.863 1.441 4.824 0.0002837
#> 
#> $type
#> [1] "plackett"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

### Clayton-Oakes
fitco2 <- survival.twostage(mph,data=diabetes,theta=0.0,detail=0,
                 clusters=diabetes$id,var.link=1,model="clayton.oakes")
summary(fitco2)

#> Dependence parameter for Clayton-Oakes model
#> Variance of Gamma distributed random effects 
#> With log-link 

#> $estimates
#>               log-Coef.       SE          z     P-val Kendall tau         SE
#> dependence1 -0.04849523 0.328524 -0.1476155 0.8826462   0.3226451 0.07179736
#> 
#> $vargam
#>             Estimate Std.Err   2.5% 97.5%  P-value
#> dependence1   0.9527   0.313 0.3392 1.566 0.002335
#> 
#> $type
#> [1] "clayton.oakes"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

fitco3 <- survival.twostage(margph,data=diabetes,theta=1.0,detail=0,
                 clusters=diabetes$id,var.link=0,model="clayton.oakes")
summary(fitco3)

#> Dependence parameter for Clayton-Oakes model
#> Variance of Gamma distributed random effects 

#> $estimates
#>                 Coef.        SE        z       P-val Kendall tau         SE
#> dependence1 0.9526619 0.3129723 3.043917 0.002335193   0.3226451 0.07179736
#> 
#> $type
#> [1] "clayton.oakes"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

### without covariates but with stratafied
marg <- phreg(Surv(time,status)~+strata(treat)+cluster(id),data=diabetes)
fitpa <- survival.twostage(marg,data=diabetes,theta=1.0,
                clusters=diabetes$id,model="clayton.oakes")
summary(fitpa)

#> Dependence parameter for Clayton-Oakes model
#> Variance of Gamma distributed random effects 
#> With log-link 

#> $estimates
#>               log-Coef.        SE          z     P-val Kendall tau         SE
#> dependence1 -0.05683996 0.3279956 -0.1732949 0.8624196   0.3208241 0.07146893
#> 
#> $vargam
#>             Estimate Std.Err   2.5% 97.5%  P-value
#> dependence1   0.9447  0.3099 0.3374 1.552 0.002297
#> 
#> $type
#> [1] "clayton.oakes"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

fitcoa <- survival.twostage(marg,data=diabetes,theta=1.0,clusters=diabetes$id,
                 model="clayton.oakes")
summary(fitcoa)

#> Dependence parameter for Clayton-Oakes model
#> Variance of Gamma distributed random effects 
#> With log-link 

#> $estimates
#>               log-Coef.        SE          z     P-val Kendall tau         SE
#> dependence1 -0.05683996 0.3279956 -0.1732949 0.8624196   0.3208241 0.07146893
#> 
#> $vargam
#>             Estimate Std.Err   2.5% 97.5%  P-value
#> dependence1   0.9447  0.3099 0.3374 1.552 0.002297
#> 
#> $type
#> [1] "clayton.oakes"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

### Piecewise constant cross hazards ratio modelling
########################################################

d <- subset(simClaytonOakes(2000,2,0.5,0,stoptime=2,left=0),!truncated)
udp <- piecewise.twostage(c(0,0.5,2),data=d,score.method="optimize",
                          id="cluster",timevar="time",
                          status="status",model="clayton.oakes",silent=0)

#> Data-set  1 out of  4 

#>   Number of joint events: 519 of  2000

#> Data-set  2 out of  4 

#>   Number of joint events: 249 of  1223

#> Data-set  3 out of  4 

#>   Number of joint events: 254 of  1226

#> Data-set  4 out of  4 

#>   Number of joint events: 615 of  968

summary(udp)

#> [1] 1
#> Dependence parameter for Clayton-Oakes model 
#> Score of log-likelihood for parameter estimates (too large?)
#>              0 - 0.5     0.5 - 2
#> 0 - 0.5  0.001223984 0.001537040
#> 0.5 - 2 -0.002090501 0.002522105
#> 
#> 
#> log-coefficient for dependence parameter (SE) 
#>          0 - 0.5        0.5 - 2      
#> 0 - 0.5  0.743 (0.068)  0.659 (0.102)
#> 0.5 - 2  0.611 (0.106)  0.746 (0.059)
#> 
#> Kendall's tau (SE) 
#>          0 - 0.5        0.5 - 2      
#> 0 - 0.5  0.512 (0.017)  0.492 (0.025)
#> 0.5 - 2  0.479 (0.026)  0.513 (0.015)
#> 

 ## Reduce Ex.Timings
### Same model using the strata option, a bit slower
########################################################
## makes the survival pieces for different areas in the plane
##ud1=surv.boxarea(c(0,0),c(0.5,0.5),data=d,id="cluster",timevar="time",status="status")
##ud2=surv.boxarea(c(0,0.5),c(0.5,2),data=d,id="cluster",timevar="time",status="status")
##ud3=surv.boxarea(c(0.5,0),c(2,0.5),data=d,id="cluster",timevar="time",status="status")
##ud4=surv.boxarea(c(0.5,0.5),c(2,2),data=d,id="cluster",timevar="time",status="status")

## everything done in one call
ud <- piecewise.data(c(0,0.5,2),data=d,timevar="time",status="status",id="cluster")
ud$strata <- factor(ud$strata);
ud$intstrata <- factor(ud$intstrata)

## makes strata specific id variable to identify pairs within strata
## se's computed based on the id variable across strata "cluster"
ud$idstrata <- ud$id+(as.numeric(ud$strata)-1)*2000

marg2 <- aalen(Surv(boxtime,status)~-1+factor(num):factor(intstrata),
               data=ud,n.sim=0,robust=0)
tdes <- model.matrix(~-1+factor(strata),data=ud)
fitp2 <- survival.twostage(marg2,data=ud,se.clusters=ud$cluster,clusters=ud$idstrata,
                model="clayton.oakes",theta.des=tdes,step=0.5)
summary(fitp2)

#> Dependence parameter for Clayton-Oakes model
#> Variance of Gamma distributed random effects 
#> With log-link 

#> $estimates
#>                           log-Coef.         SE         z        P-val
#> factor(strata)0-0.5,0-0.5 0.9121858 0.06870365 13.277108 0.0000000000
#> factor(strata)0-0.5,0.5-2 0.4345982 0.11298180  3.846621 0.0001197581
#> factor(strata)0.5-2,0-0.5 0.3910945 0.11496308  3.401914 0.0006691581
#> factor(strata)0.5-2,0.5-2 0.6844110 0.05891533 11.616858 0.0000000000
#>                           Kendall tau         SE
#> factor(strata)0-0.5,0-0.5   0.5545418 0.01697153
#> factor(strata)0-0.5,0.5-2   0.4357204 0.02777863
#> factor(strata)0.5-2,0-0.5   0.4250558 0.02809506
#> factor(strata)0.5-2,0.5-2   0.4978160 0.01472855
#> 
#> $vargam
#>                           Estimate Std.Err  2.5% 97.5%   P-value
#> factor(strata)0-0.5,0-0.5    2.490  0.1711 2.154 2.825 5.407e-48
#> factor(strata)0-0.5,0.5-2    1.544  0.1745 1.202 1.886 8.675e-19
#> factor(strata)0.5-2,0-0.5    1.479  0.1700 1.145 1.812 3.365e-18
#> factor(strata)0.5-2,0.5-2    1.983  0.1168 1.754 2.212 1.290e-64
#> 
#> $type
#> [1] "clayton.oakes"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

### now fitting the model with symmetry, i.e. strata 2 and 3 same effect
ud$stratas <- ud$strata;
ud$stratas[ud$strata=="0.5-2,0-0.5"] <- "0-0.5,0.5-2"
tdes2 <- model.matrix(~-1+factor(stratas),data=ud)
fitp3 <- survival.twostage(marg2,data=ud,clusters=ud$idstrata,se.cluster=ud$cluster,
                model="clayton.oakes",theta.des=tdes2,step=0.5)
summary(fitp3)

#> Dependence parameter for Clayton-Oakes model
#> Variance of Gamma distributed random effects 
#> With log-link 

#> $estimates
#>                            log-Coef.         SE         z        P-val
#> factor(stratas)0-0.5,0-0.5 0.9121858 0.06870365 13.277108 0.000000e+00
#> factor(stratas)0-0.5,0.5-2 0.4125895 0.09030649  4.568769 4.905978e-06
#> factor(stratas)0.5-2,0.5-2 0.6844110 0.05891533 11.616858 0.000000e+00
#>                            Kendall tau         SE
#> factor(stratas)0-0.5,0-0.5   0.5545418 0.01697153
#> factor(stratas)0-0.5,0.5-2   0.4303170 0.02213812
#> factor(stratas)0.5-2,0.5-2   0.4978160 0.01472855
#> 
#> $vargam
#>                             Estimate Std.Err  2.5% 97.5%   P-value
#> factor(stratas)0-0.5,....      2.490  0.1711 2.154 2.825 5.407e-48
#> factor(stratas)0-0.5,.....1    1.511  0.1364 1.243 1.778 1.689e-28
#> factor(stratas)0.5-2,....      1.983  0.1168 1.754 2.212 1.290e-64
#> 
#> $type
#> [1] "clayton.oakes"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

### same model using strata option, a bit slower
fitp4 <- survival.twostage(marg2,data=ud,clusters=ud$cluster,se.cluster=ud$cluster,
                model="clayton.oakes",theta.des=tdes2,step=0.5,strata=ud$strata)
summary(fitp4)

#> Dependence parameter for Clayton-Oakes model
#> Variance of Gamma distributed random effects 
#> With log-link 

#> $estimates
#>                            log-Coef.         SE         z        P-val
#> factor(stratas)0-0.5,0-0.5 0.9121858 0.06870365 13.277108 0.000000e+00
#> factor(stratas)0-0.5,0.5-2 0.4125895 0.09030649  4.568769 4.905978e-06
#> factor(stratas)0.5-2,0.5-2 0.6844110 0.05891533 11.616858 0.000000e+00
#>                            Kendall tau         SE
#> factor(stratas)0-0.5,0-0.5   0.5545418 0.01697153
#> factor(stratas)0-0.5,0.5-2   0.4303170 0.02213812
#> factor(stratas)0.5-2,0.5-2   0.4978160 0.01472855
#> 
#> $vargam
#>                             Estimate Std.Err  2.5% 97.5%   P-value
#> factor(stratas)0-0.5,....      2.490  0.1711 2.154 2.825 5.407e-48
#> factor(stratas)0-0.5,.....1    1.511  0.1364 1.243 1.778 1.689e-28
#> factor(stratas)0.5-2,....      1.983  0.1168 1.754 2.212 1.290e-64
#> 
#> $type
#> [1] "clayton.oakes"
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"

 ## Reduce Ex.Timings
### structured random effects model additive gamma ACE
### simulate structured two-stage additive gamma ACE model
data <- simClaytonOakes.twin.ace(4000,2,1,0,3)
out <- twin.polygen.design(data,id="cluster")
pardes <- out$pardes
pardes

#>      [,1] [,2]
#> [1,]  1.0    0
#> [2,]  0.5    0
#> [3,]  0.5    0
#> [4,]  0.5    0
#> [5,]  0.0    1

des.rv <- out$des.rv
head(des.rv)

#>   MZ DZ DZns1 DZns2 env
#> 1  1  0     0     0   1
#> 2  1  0     0     0   1
#> 3  1  0     0     0   1
#> 4  1  0     0     0   1
#> 5  1  0     0     0   1
#> 6  1  0     0     0   1

aa <- phreg(Surv(time,status)~x+cluster(cluster),data=data,robust=0)
ts <- survival.twostage(aa,data=data,clusters=data$cluster,detail=0,
               theta=c(2,1),var.link=0,step=0.5,
               random.design=des.rv,theta.des=pardes)
summary(ts)

#> Dependence parameter for Clayton-Oakes model
#> Variance of Gamma distributed random effects 

#> $estimates
#>                 Coef.        SE         z        P-val Kendall tau         SE
#> dependence1 2.0850765 0.1493742 13.958742 0.000000e+00   0.5104131 0.01790216
#> dependence2 0.9578921 0.1170992  8.180176 2.220446e-16   0.3238428 0.02676821
#> 
#> $type
#> [1] "clayton.oakes"
#> 
#> $h
#>             Estimate Std.Err   2.5%  97.5%   P-value
#> dependence1   0.6852  0.0389 0.6090 0.7614 1.835e-69
#> dependence2   0.3148  0.0389 0.2386 0.3910 5.813e-16
#> 
#> $vare
#> NULL
#> 
#> $vartot
#>    Estimate Std.Err  2.5% 97.5%    P-value
#> p1    3.043  0.1057 2.836  3.25 3.537e-182
#> 
#> attr(,"class")
#> [1] "summary.mets.twostage"