G-estimator for Cox and Fine-Gray model

Computes G-estimator $$ \hat S(t,A=a) = n^{-1} \sum_i \hat S(t,A=a,Z_i) $$ for the Cox model based on phreg og the Fine-Gray model based on the cifreg function. Assumes that the first covariate is $A$ and this is a factor. Gives influence functions of these risk estimates and SE's are based on these.

survivalG(x, data, time = NULL)

Arguments

x

phreg or cifreg object

data

data frame for risk averaging

time

for estimate

Author

Thomas Scheike

Examples

data(bmt); bmt$time <- bmt$time+runif(408)*0.001
bmt$event <- (bmt$cause!=0)*1
dfactor(bmt) <- tcell~tcell

fg1 <- cifreg(Event(time,cause)~tcell+platelet+age,bmt,cause=1,
              cox.prep=TRUE,propodds=NULL)
summary(survivalG(fg1,bmt,50))
#> risk:
#>       Estimate Std.Err   2.5%  97.5%   P-value
#> risk0   0.4332 0.02750 0.3793 0.4871 6.415e-56
#> risk1   0.2727 0.05863 0.1578 0.3876 3.307e-06
#> 
#> Average Treatment effects (G-estimator) :
#>    Estimate Std.Err   2.5%  97.5% P-value
#> p1  -0.1605 0.06352 -0.285 -0.036 0.01152
#> 
#> Average Treatment effect ratio (G-estimator) :
#>       Estimate   Std.Err      2.5%     97.5%     P-value
#> [p1] 0.6294718 0.1392308 0.3565845 0.9023592 0.007785081
#> 

ss <- phreg(Surv(time,event)~tcell+platelet+age,bmt) 
summary(survivalG(ss,bmt,50))
#> risk:
#>       Estimate Std.Err   2.5%  97.5%   P-value
#> risk0   0.3461 0.02709 0.2931 0.3992 2.132e-37
#> risk1   0.4360 0.05972 0.3189 0.5530 2.870e-13
#> 
#> Average Treatment effects (G-estimator) :
#>    Estimate Std.Err     2.5%  97.5% P-value
#> p1  0.08984 0.06292 -0.03348 0.2132  0.1533
#> 
#> Average Treatment effect ratio (G-estimator) :
#>      Estimate   Std.Err      2.5%    97.5%   P-value
#> [p1] 1.259545 0.1894615 0.8882069 1.630882 0.1707161
#>